(-18.390 [tm] action: graph sync (first half) (.. 4386))
(we set out to “fix” syncing without having any deep understanding of
what it was or how it worked. (yes, it’s been that long.) once we
achieved that, we realized that its regression order slates it for
after almost every other action; however at that point we had done
significant enough work that the cost of stashing this work in a branch
would be significant. so here it is, half finished.)

(2 in 18.69, 17.95)