(-18.380 [tm] action: association add (!) (.. 4471))
Discussion: our enthusiasm stems from the idea that this is the
  “top” of the mountain: adding associations is like the essential
  operation of the subject application; and we hope that it is
  mostly downhill from here.

Re-new feature:

We don’t want to rewrite the file when there was no change in the
file (let’s pretend that writes are costly, which they are).
in the old way we did something kind of nasty where we skimmed for
events and read whether the event reported that a change happened.
now, experimentally, we have a dedicated reader method on entities
just for this purpose..

Fun tangent:
 we spiked the theory without a name at [#hu-066]

(times)

  11:49         begin create
  14:16  (147)  finished first case (about IO)
  18:36  (260)  after lots of cleanup and history rewriting,
                we are on to the next case which is minus 9
  18:51   (15)  8 left
  19:48   (49)  seven left
  19:54    (6)  six left, nap
  00:01  (247)  restarted
  00:18   (17)  five left
  01:42   (84)  finished flowchart
  02:57   (75)  making the flowchart was a GREAT idea.
                code "feels" "flawless" now. four left
  03:21         break
  04:23         finished the one that needed the o.b - three left
  04:47   (24)  restored attributes etc without too much pain. two left
  04:52    (5)  yum one left
  05:11   (17)  whole file!!!! (very rough)

(2 in 17.71)