(-17.991 [cu] action: survey edit upstream 3/3 unset (.. 4508))
Discussion: the way we used to accomplish this particular
“unsetting a singleton component” (in this case the upstream)
was something of a hack for lack of good design: passing the empty
string as the upstream identifier had special meaning that amounted
to unsetting an existing upstream. fortunately the “design consequence”
of this design objective was fairly straightforward and painless:

(quick parenthetical: the idea of a “singleton association” per se
(in these words) was probably coined in this commit.)

for all singleton associations that we want to be able to unset in
this way, we simply add a flag parameter to our edit action with a
name in the form of `unset_<<association>>` (so `unset_upstream`
here). then in the performer, we detect this special verb “unset” and
act accordingly.

As a side effect of some consequence, we discovered that although
we were parsing our existing config file on read, we were not
unmarshmaling it. (however this used to work in the past, we don’t
care, it was unreadable and had no value; case in point next paragraph..)

as a consequence of *this*, then, tests that once passed with various
degrees of garbage for their upstream paths and garbage otherwise now
break. for some of these tests you will see we changed their upstream
paths to be existent paths. in the past, the system would allow garbage
to come in in this manner (perhaps anticipating that it would be
overwritten) but it is not acceptable now.

  06:46  begin again
  02:21  started flowchart for unmarshaling
  14:15  no more '.path' - only '.survey_path' (too confusing otherwise)
         one thing we did is we accept on blind faith the integrity
         between the survey entity  and the document -
         at write time the .. um...

(2 in 23.24, 21.86)