(-17.985 [arc] spike a new kind of model (.. 4528))
Discussion: classic [arc] feels too heavy. [fi] “attributes” feel too
old, dusty, and in places clunky. (also still feels like too pointed
a dependency.) what we do in the re-finished legacies ([sg], [tm], some
[cu]) seems hodgepodge and/or custom and/or in development.

so, for today what we needed was a middle ground model model, one that
builds off of the simplicity of SimpleModel but adds basic type
assertion and tries to lay you on a path that puts an upgrade path
ahead of you (but still on the same path, to insist on the metaphor).

(times)
  02:57  decide its time to rename to arc
  08:42  finished [arc] rearch
  10:13  begin new kind of model
  04:00  finish new kind of model

(2 in 26.17)