(-17.983 -7 (renames and) begin upgrade from ruby 2.2.3 to 2.4.1)
Scope of this cluster:

rubygems.org changed its SSL certificates to use a stronger form of
encryption, which we ran into while trying to install a gem for a
different project.

we couldn't get our older version of ruby to find these certifcates so
we simply blew it all away and re-installed latest stable ruby. this
(of course) tripped off a good cascade of maintence work for us, namely:

  - refactor those parts of our code that use ruby language/
    runtime features that are deprecated or no longer supported.

  - refactor code that relies on stdlib API details that have changed

  - refactor code to play nice with latest stable gems of our small
    set of external gems.

the N commits in this cluster are towards this end, but also they
drew our attention to one glaring issue: the installation scripts
were is shoddy shape. so the bulk of the interesting work is towards
the end of improving the five or so installation-reltaed scripts in
[sli].

As for this commit:

  - give these initial installer scripts a bigger numberspace so
    we have room to add many more files in between them (hopefully
    we won't need this)

  - task has been renamed to "installed gem" (from "gem file builds")
    to indicate that its passing state is simply the gem having been
    installed. it is not the case that the task must install the
    gem itself: if the gem is already installed with the correct
    version, this task is a no-op.

 - "build and install gem" became "install essential gems"

 - "gemspec" became "install remaining gemes"

 - "devify" is not a good use case for a neologism.
   it is best just to call this "symlink" here, because that's
   exactly what's happpening.