(-17.973 -1.12 add line-numbers report (( 19 )) )
the main thing here is that the new “reports” architecture has “rocked”
the existing stuff such that the original (largely stubby) stuff has
been “subjugated” to be just another report.

see the amusing “bittersweet creation myth” in the comment of the
line numbers report.

we decided that it’s not worth it to aggregate the multiple files
of file diffs (and do the clever trick with shortening the paths);
in fact it may be a misfeature.

we added descriptions to the

no tests yet! the “line numbers” report works against the magnetic
(of the main collection not the specialized collection) with the
longest name. the stub that displays diffs should work against
arbitrary file names, because it doesn’t parse anything yet.

    07-19 10:18  zzz
          15:05  wake, start
    07-23 21:36  resume - going thru every file in [bs] tryna getta sense
          22:53  reached end of first nontrivial file
    07-25 19:08  now
.