(-17.972 -1.11 add coverage for the 2 reports and.. (( 24 )) )
now that we feel we are getting past a “prototyping” or a “proof of
concept” phase,

finally start to add coverage. the “echo file paths” report amounts to
a “hello world” or a “ping” just to see if the very basics are working.

not covered, added 2 “help screen” operations: one to display a summary
of every report, and two to display the full description of any one
report.

    07-25 23:41  begin working on files names list coverage
          01:39  finished 2 kinds of help screens
          02:26  covered file path names
          03:00  zzz
          10:27  wake
          12:54  done with test for 2 reports