(-17.970 -1.9.15 interrupt everything.. ( 47 )) )
interrupt everything to use ‘parser’ gem instead:

so, we didn’t do due diligence on researching tech. we went with the
first ruby gem we found (“ruby_parser”) because we mistakenly thought
it was the “main” one (based solely on the fact that it supported
parsing latest stable ruby). it would turn out that many of the features
we thought we were going to add (eventually) to a fork of the project
(a rewrite, really) exist (form the sounds of things) exactly like we
would want them to.

because this will be a semi-major architectural overhaul, we’re just
gonna freeze a snapshot of everything we have here as-is (halfway
finished on this report), knowing that most of it may change (indeed
a lot of it here was the beginnings of a crutch, working around the
shortcomings we encountered of our tech).

  - added this “after all files” hook as a place to flush comprehensive
    statistics summary

  07-27 01:31  drive stuff forward by making preview selected report
        11:25  slept at 4 am. woke. playing with some stuff
  07-28 17:00  everything has changed - we don't know why we
               accidentally started using 'ruby_parser' (probably
               familiarity) but we are going to change radically
               everything so we can use 'parser' instead.