(-17.961 -1.9.6 (C) rough version of FULL grammar..)
wahoo we think this is a proof of concept that we can get
granular error reporting.

woah this is a rough version of THE FULL grammar
(error reporting is incomplete). also added build script.

(at this point we had looked at just about every example
(all in C) in the ragel guide.)