(-17.929 +1.2.2 wean API ping (( 115 )) )
spike lots of code that we will want to abstract away later, probably.

it was optimal to add a new file to cover a purpose-built testpoint
like “ping” so that we have a dedicated “silo” whose only purpose is
to demonstrate an idealized action (sadly, no parameters) and an
idealized way to test it.

  - [ze] sneak in one case for “express result”

(times)

    08-24 19:09  start to implement and cover API calls
          22:16  passes tests