(-17.926 +1.2.5 wean CLI deliterate (.. 4655))
Discussion: the target action that we are restoring requires the
`par` method of the legacy expression agents. our new CLI is hard-coded
to default to our newer, simpler expags. our reasonable answer to this
is to want to be able to specify any arbitrary expression agent per
action.

so, new in this work, we wanted “action-specific expression-agent
injection” to be as easy as injecting associations. (in the past, the
application’s CLI client could provide its own custom expag; but
if there was a way to get an action-specific expag, however it was
done it was not as easy as this.)

to this end we re-introduced a new “resources” and some smell cropped
up. (we don’t want CLI to subclass resources, but we are avoiding
composition too, because it seems absurd for CLI not to hold the
main resources directly, so as it stands there is some compositional
redundancy between CLI and resources.)

  - we don’t love putting `Modalities` everywhere. need to think
    about this. (we could do a nasty filesystem peek but we hate that)

  - [ts] allow the CLI under test to accept the newer style of
    configuration where it accepts a “definition block” instead of
    an iambic list of parameters.

  - [ts] when running a single test, accept shorthand “-12:34” as well
    as “-12-34” (to mean “any tests on lines 12-34 (inclusive)”).

  - [ze] construct legacy (proc-based) expags in definition block.

(2 in 10.58)