(-17.919 +1.2.12 reinvent stack. birth of “steps” (.. 4675))
this commit is concerned with two related .. concerns. I) the stack
mess and II) the hook mess.

I. the stack mess

much like in our previous works, an interface “stack” crept into this
machinery organically and progressively. we didn’t realize what aspect
should and shouldn’t be part of the stack frames until we got to the
end of integrating it. for example, we didn’t realize that we would
want to associate expression agents (which you can now inject per-
operation per-modality) with stack frames until we finished.

II. the hooks mess

in effort to be DRY between API and CLI we ended up with a super
convoluted hook interface. first, some background:

  a typical operation needs to be constructed with the “invocation
  resources” object as its only argument.

  part of the “invocation resources” is the expression agent.

  in more complex legacy apps, the expression agent is built custom
  for the operation (so it can have the operation inside of it, so
  it can do reflection of the operation’s associations.)

note this creates a circular dependency between the operation and
the expression agent. nominally we try to avoid mutability so our
current convoluted way around this is to build the operation first
with `allocate`, then build the expag around the operation, then
`initialize` the operation with the expag.

to boot, the above convoluted construction for operations under CLI
needs to happen in conjunction with “injection” (customization) of
ad-hoc expression agents for any particular options.

what’s more, keep in mind the whole (new) API stack existed already
when we came along with this architecture. we tried not to disrupt
it too much while being DRY, and in so doing disrupted it a lot with
our “hooks” interface.

this commit largely cleans up the mess we made there.

Small things

  - symbols used as if they are consts in C that look `_like_this_pl_`
    are now named to look `LIKE_THIS_pl` - this if they “belong” to
    one sidesystem but are used across several. marking the end of the
    symbol name with a sidesystem sigil gives you some recourse if you
    want to look for documentation.

(original hooks gripe)

the hooks interface was convoluted. there was no rigidity to the hooks
being called in any order or with any guarantee that they would be
called. rather, make the remote library more friendly to a stacking
client by resulting in the (intermediate) "step" object that itself
can be used to create a bound call..

(times)
    09-08 01:54    very beginning of trying to add stacks to new CLI
    09-09 14:16    original hooks gripe above
    09-12 05:53    almost done / now
          17:40    universe integration

(2 in 12.04, 9.87)