(-18.859 -0.40 de-monolithize tests (( 289 )) )
commentary: due to a confluence of these factors:
  - the new strictness that comes (necessarily) with the
    correction of our escaping mechanism
  - the shapes of code (features) that are being unparsed now that
    weren’t unparsed before
we are finding that we will probably need to cover a lot more things.
having a monolith test file was untenable and likewise the monolith
unparsing magnet was ugly too. (we began breaking it up in the last
commit.) anyway, to the end of flooding our timeline with lots more
coverage, we are splintering out things into categories that accord
with the classes of location map..

  - rename, commit everything as-is (-41)
  - subdivide guys into location map classes ordered (-40)