( 12.5  magnetic for commands via directory (41))
all we’re trying to do is get to a hello-world that parses parameters,
so we can get to boogey-ing. but to do that it looks like we will need
at least 2 more interceding magnets. (one: this one. two: something
about parameters. (not necessary for hello world but necessary for
general use.))

we revisit a central algorithm we’ve used elsewhere, that of parsing
requests for endpoint trees in the obvious way. this is laying down
some overly-optimistic tracks for the future, where we have deep
endpoint trees. (but deep trees are certainly not covered here).

and:

  - (lots of small edits to existing docs, after being able to review
     them by seeing them in github for the first time. there has to
     be a better way.)

  - the newer, more robust algorithm refactors our main CLI file
    in a way that (along the way) sees it try to break out into
    small functions rather thank medium-sized class monoliths.

  - introduce a here not covered and unresolved issue where we’re
    gonna have to translate and un-translate names `like_this` to
    and from names `like-this`

(0.042s)