(+grep-dump.09.08 add empty Pipfile…)
Synopsis: the main purpose of this commit is to capture a snapshot
of the Pipfile with nothing in it. this way, as we add dependencies
to this project, we can track exactly when and why we added them…

Background: originally when we started dumping python work, we knew
that there was a “right way” to handle versioning our dependencies;
we just didn’t know what the right way was at the time. now, we
*think* we know the right way to handle this (sort of — we’re not
sure how we want to handle sub-project-specific dependencies yet,
but later for that).

Details:

  - spike a heavy doc about this.