(+grep-dump.09.09 stop server cleanly (12))
Discussion: the largish comment committed in the giant code blob
explains this: we were debugging an issue where when the server shut
down it failed to clean up its resources (not our fault!). (there is
a chance this fixed in the jump from 3.5 to 3.6 which happened to us
while we were working on this!)

so anyway we needed to dissect some of the server startup code to
expose all of its inner scopes so we could access things if necessary.

once we get a better sense for what’s going on, we’ll get rid of this
mess later; but we want it for reference when we chop and screw it
later..

Also:

  - (tiny edits to body copy of [#018]. (“polish” added after reading
    it published to github.))