(+slack-files-bot.05.08 tunneling and verification (31,8))
This is a bit of a mini-milestone, because it’s the first time our app
actually talks to slack using its own callback webserver. we had to
take some time to get our head around all its little moving parts, and
in so doing went on several important tangents (accounted for in the
immediately previous commits) to grant us the reproducibility and
control we wanted to go forward with this stuff.

But we now have:

  - start up scripts for both test environment and dev environment.
    (we can run both servers at the same time)

  - tunneling, so verification

In this commit:

  - add one more element to the expected list of environment variables
    (for verification)

  - started document on ngrok because had OCD and anxiety.
    it “paid off” because the verification at first didn’t work
    then it did, because of our OCD:

  - having parsimonious but extant logging (tracing) saved our ass
    when the verification at first didn’t work.

  - broke out stuff from main README into what became a document on
    deep musings about security and environment variables

  - added a README to the “700” tree of tests only so that we can
    say that thing about wanting to rename the one file.

  - theme music: Herzogenberg - Piano quartet in E Minor,
    Op. 75/Andante - Belcanto Strings - Andreas Frolich, Piano

(0.013s)