(+slack-files-bot.05.09 pipenv install celery)
Synopsis: (the next commit will explain the visual test justifying this)
  here we just want an isolated commit with the new dependencies,
  and the spike of docs.

this contains a rabbitmq but we may end up not using it. we include it
here because A) we were able to get it to “work” (for some definition
of) and B) we had documented its usage etc, so we wanted to go ahead
and fold that in now.

  2018-03-31 10:28  start reading about OAuth permission scope
       04-01 01:57  start reading OAuth RFC
             13:43  thinking about threads
             13:57  found the thing in slack (ref'd in doc) re: threads
             14:12  gonna read about celery
       04-02 13:41  install redis