(+slack-files-bot.05.10 visual: task queue demo)
Synopsis: use Miguel Grinberg's excellent (if one year old) tutorial
of a progress indicator as a visual demonstration of our task queue
stack working.

Objective & Scope of this commit: show the task queue working for
asynchronous handling of long jobs.

Discussion:

this gets us one step past a "hello world" demo of a task queue
because the client keeps pinging it while it is in progress until it
completes, giving progressive updates (visually).

we are not interested in the fool's errand of trying to cover all
this reliably. instead, we will enter into a small-step phase of
visual tests. meh.

Regression notes:

  - have celery and rabbitmq installed
    as described in the docs in this commit.

then in three terminals (we like them right to left):

  - in one terminal: `cd Z/redis-stable && ./src/redis-server`

  - in another terminal:
         celery -A upload_bot._magnetics.webserver_via_behavior.celery \
           worker --loglevel=info

  - in a THIRD terminal `python3 -W 'default::::' upload_bot/run.py`

then with your webserver (flask) now running, go to `localhost:5000/wahoo`
and press the button. a progress indicator

Brief narrative:

we got to the point where we were receiving events from slack in a
recognizable way (JSON). the next step was to learn about responding to
these events asynchronously (as the slack docs recommended). it took us
a bit of re-reading but we found where the docs recommended this.
(perhaps we should have taken better reading notes the first time but
meh.)

so we concluded that we wanted celery and rabbitmq (would later scale
this back to just redis for now). the first major snafu hit when
we had our webserver process blow up because of a deprecation warning
that would hit in celery and we had warnings set to be treated as
errors.

the way around this (but to still satisfy our OCD of wanting warnings)
is to run the webserver with (again):

    python3 -W 'default::::' upload_bot/run.py

rather than what our doc perscribed before this commit.

Some Details:

  - spike a bunch of docs: celery, redis, rabbitmq

  - main flask file gets spiked with a bunch of cruft
    we don’t want to stay there

  - start a templates folder, OK fine whatever

(times)

  2018-03-31 10:28  start reading about OAuth permission scope
       04-01 01:57  start reading OAuth RFC
             13:43  thinking about threads
             13:57  found the thing in slack (ref'd in doc) re: threads
             14:12  gonna read about celery
       04-02 13:41  install redis
       04-04 05:03  finally saw progressive demo working
(.)