(+sakin-agac.09.20 explode the “script lib” monolith! (172))
Discussion:
the “script lib” previously confused ambition with aggression. that is,
it was *aggressively* ambitious. what is (of course) better is a
“toolkit” approach, but these take more effort to design well.

now, the library user need not feel as overwhelmed: the library is
now more modular and graded. rather than one monolith, the library
is now something like five very small pieces. in this arrangement,
the library user may choose to employ the higher-level pieces or maybe
only the lowest level ones as appropriate.

she can now opt-in at a level of abstraction that is the best fit for
her needs, perhaps using only the lowest level magnetics. she can
scale-up to higher and higher levels of grade as is useful.

as always this is of course experimental and is certainly subject to
more change as it gets more real world use.