(+sakin-agac.09.30 (3  4) snapshot: statement of purpose (46))
Discussion: here we add the trivial `to_line` for “near” items, and
bring ourselves up to the edge of the cliff to what we have to work
on next. That is, if you "unwip" the "wipped" test, you will see the
exact feature that we do not have that we want to have (sort of a
rule-of-three's of TDD thing); that we intent to implement next.

  - we learn that python’s `unittest` has no good tagging mechanism
    (e.g. to “wip” tests)