(+sakin-agac.09.39 (3  2) refine: variable endcaps, unify schema (78))
Discussion: when applied to our practical case, a problem became
apparent around endcaps: some of our real rows in the document had
“endcaps” (a trailing final pipe before the final newline on a line),
and others did not. the presence or absence of these endcaps (indeed,
the question of whether or not they were parsed as an endcap or rather
such a feature was parsed as indicating an empty cel); this led to
undesired behavior around counting cels and what happened during
row-level merges.

Also: the idea of the table “schema” was kind of half-baked and half
realized and manifested in various ways at various points in the code.
(there were at least two class-likes in the code with the word “schema”
as part the name (or something near this).)

Our new imagined spec is something like this:
  - the first and second schema row (so, all of them)
    must have an "endcap".
  - on subsequent rows, the "endcap" is optional
  - subsequent rows can have less (not more) cels than the schema

(in practice, we have not covered the requirement that the second
schema row have an endcap; and we suspect that this behavior is not
present either.)

Additional Commentary:
  - make how we validate cel count (on second schema row, on business
    rows) more abstracted, to flex-to-policy better

(times - yowza)
  05-14 17:00  start cover mostly already working procure thing
  05-16 08:20  begin
  05-19 17:44  try again - change cel parsing life - row parsing lyfe
  05-25 12:25  finished this