(+sakin-agac.09.43 (3 -8) spike “listener via expectations” (101))
..and strange format adapter name

Discussion: we set out to cover only this one behavior point (when
a strange format adapter name is used); but we also had this underlying
motivation:

to get “over the wall” of modality-specific tests. you see, we like the
idea of writing our applications as “moralistically pure” as is
reasonable. what this means may have to be the subject of another
[..]

but how it pertains to this work is that we want to get away from
testing our essential functions in terms of one or another particular
modality. for our immediate purposes, what this means is testing
“emissions” rather than modality-tailored expression of those
emissions.

in order to do *that* we spiked a nifty new DSL (sketch).

(times)
  05-28 04:07  begin in earnest bad FA name
        05:59  conceive of crazy new DSL for mock listener
        06:58  pretty good start zzz
        15:36  looking good