(+sakin-agac.09.47 (2   ) snapshot: capture endcap issue (117))
Discussion: we are capturing an undesired behavior. (the test is written
as if it is the desired behavior). The undesired behavior is that if a
near row is being “updated” in a sync and it does *not* at first have
an “endcap” and the “prototype row” *does* have an endcap; the update
“adds” an endcap to the inserted row.

the desired behavior is that the particular “encappiness” of a row is
preserved when it is updated in a sync.

Additionally, we cover a bit of an edge case which is: our state machine
traversal thing needs to do the right then when the markdown table is
the last thing in the document.

(times)
  06-05 11:29  begin attempt at a successful sync
        11:56  it looks it worked but endcap thing
        13:42  now