(-tag-lyfe.19.2.4 markdown table as data producer (167))
Long Narrative: the ultimate ostensible objective is to have our
publishing platform (blog-like, CMS-like) on our own specific terms
(static site generator-like, heavily git based, comfortable running
locally, able to deploy on heroku).

towards this we want to pare down the existing options available on
heroku. reasonably we want to do the craziness of parsing tag queries
and doing a map-reduce on our tagged set of data.

so we at first thought we might actually pipe lines of text across
platforms to our old ruby work (“snag” ([sg]) is largely such a system)
so we wanted this `stream` script.

but also, the main objective (locally) was to study the space of parser
generators that output python. thinking forward to how a
`filter_by_tags.py` script would eventually work (in terms of its real
interface) we realized that would not (at first) be actually piping
yet, so the “stream” script was deemed as maybe superfluous.

so we stashed the sizable work we had begun of making markdown tables
streamable (mostly just glue, because syncing is a functional superset)
so that we could focus on our survey of parser generators..

but then (thrillingly) as we were making scrapers to gather the domain
data it occurred to us: why are we scraping html to create markdown
when the source (far) data is itself markdown natively? (it’s sort of
a philosophical question: does it mean anything for data to “live”
somewhere if it’s replicated isomorphically in different forms in
different places?)

so anyway one option is to try and achieve isomorphism: links that
started as markdown (`[foo](www.bar/baz)`) come to us as html anchor
tags. we would then have to parse them back int a sort of simple
structured AST-like (really just plain old programming) and in effect
achieve a tiny sliver of (html2markdown - wow). but the other option
(worth trying, at least) is to parse the source (far) data *as*
markdown. (several cans of worms here..)

because github (at least) allows us to see the “raw” (markdown) text
of our target far README just though plain old HTTP request, this might
be fun and not so bad. but anyway, to actually see this though we would
need to be able to traverse markdown tables as a “first class” data
producer, something we wanted to do anyway at the start of this story
(just for the “wrong” reason). this commit realizes that first detour,
of poising a markdown table as a data producer proper.

Refactors that happened here:

  - give symmetry with the names: we have taken to liking the prefix
    “open_” for things that create context managers, and want that to
    be something of a convention. so as it happens, the API functions
    that format adapters can expose both return context managers, so
    now they follow suit and fall in line with this new convention,
    so `session_for_sync_request`->`open_sync_request`.

  - as it has worked out, deep confusion stems from the use of the
    word “item”. the platform’s use of `foo.items()` means a generator
    that produces name-value pairs in tuples. but we have a variety
    of semantics for it here. so for now, we prefer `dictionary`
    when the item is a dictionary (rather than “JSON object”
    (except for scripts)) and we add a variety of qualifiers before
    the term “item” in various places.

  - “nearstream foos” and “farstream foos” aren’t cool any more. it’s
    “near foo stream” and “far foo stream” duh.

Other musings that came to use while writing this:

  - when we say we are parsing “markdown” what we really always mean
    is “github-flavored markdown” because (as far as we know) github
    invented the markdown table. github’s publicly available markdown
    parser is (again, as far as we know) written in ruby so we might
    get to straddle platforms after all. in effect we would rewrite
    the whole format adapter so it leaps across processes and
    platforms yikes/meh

(times)
  06-18 05:14  begin looking at second document
  06-19 02:33  visual test looks good, start to lock down
        16:01  now