(+tag-lyfe.19.2.2.1 refactor to interleaving algorithm (206))
Discussion: life is journey, and when code is life, code is journey.
when we started this pulse of work, we thought we were eradicating
the old way with the new way; then halfway through we realized we
wanted to keep the old way around and just use the new way as a default.
now that we are finished we are realizing that maybe the old way should
be the default, and the new way opt-in. no biggie.

  - add quite a few tests that test for failure when collections
    aren’t ordered. (we have to hit this at different legs of the
    algorithm.)

  - the order of items changed in several tests

  - changed UI language to “key” from “human key”. (we aren’t as
    crazy about the term as we once were. we don’t want to be married
    to it too deeply. “key” is sufficient for many contexts.

  - took “dict” out of variable names, for contexts where maybe it’s
    not known or significant that it is a dict stream

(times)

  06-25 16:30  take another whack at the big collect
        16:40  (worked around another 508)
  06-26 16:20  begin quick before dance
  06-27 22:15  begin better coverage for and fuller implementation of sorting
        22:16  pause for shower
  06-29 09:50  begin interleaving
  06-29 13:21  begin implement the new way
  07-02 00:06  begin bring back old way
        03:21  begin final run down fix
        18:07  now