(-tag-lyfe.9.SNAPSHOT capture close but not quite grammar (105))
Discussion: we felt like we had the overall grammar working well,
but as it would turn out it's trickier than we at first anticipated
to match correctly the runs of "garbage" between taggings (which is
to say to avoid falsely matching not-taggings).

we captured this intermediate savepoint (that shows in its tests one
or two illustrative failing tests) to illustrate how drastically we
had to change the grammar in order to get our last few cases.

also:
  - get rid of unnecessary sub-classes
  - oops we did’t memoize the creation of the walker before

(times)
  07-28 04:05  begin deep thought about how the grammar of
               taggings will work - we have a good idea
        04:40  ok begin write some tests
        05:20  woah just read a bunch of docs about grammar features.
               begin pseudogrammar.  zz
        15:14  try again - start to play with grammar sketch
  07-29 01:38  begin write expect cases
        02:00  begin write the main magnet
        02:13  writing the wiring
        04:04  FINALLY parsing some garbage
        04:19  don't get tripped up by hashtag in the middle
        04:47  (that was hard - get the lookahead to look at two things)
        04:53  ok begin doing stuff with AST walking
        09:04  much better - created document model and
               some test run with it
        09:42  almost done zz
        16:23  all but one - a tricky edge case