(-tag-lyfe.8 double-quoted string name components (112))
Discussion: half way thru finishing this we realized it would be
prudent to have the query grammar re-use *ALL* of the tagging grammar
(as it pertains to parsing taggings, so really all of it).

Small thing: our tagging model changed (slightly) yet again: now it
has this idea of “name components” of each tagging node. also now the
“tagging” is this outermost wrapper, it is not the recursive part.
this changes how we build our native AST’s near where we expect there
to be recursion: now, a “tagging” is itself not the recursive thing,
the node is. hence you see a name change in lots of places to
`tagging_node` instead of `tagging`.

(times)

  08-01 03:12  start deep tags
        04:47  ok finished rough pass at grammar
        06:02  ok making the AST for double-quoted string
        07:07  came up with new improved tagging model
               to handle quoted strings sanely
        08:02  good job, parsing double quoted string value
               using fancy new model. happy with progress. zz
  08-02 03:53  begin - build out test support to have memoized subject
        07:06  no really let's try again
  08-03 04:22  no really let's try again
        05:06  looks like quotes in done, now regress
               back to everything else
        06:39  start last looks and notes
        06:56  now