(-tag-lyfe.4.5 case study 1: simplifying comparison key..)
the purpose of this "case study" is to capture a snapshot showing
a real-world use-case of our "syncing" to reveal both ways it is
useful and ways in which it might be improved.

the over-arching near-term objective is to get this (now several weeks
old) "dim sum table" up to date with the latest dataset from heroku.
(that's the whole point of this synchronizing empire, is to make stuff
like this easy..)

in order to re-use benefits gained from another "dim sum" (the
(gasp) multi-source syncing of items for the great parser-generator
shoot-out); we moved this "keyer function" from an ad-hoc producer
script up into a location we can reach it by several scripts.

among other things, this keyer function is better at detecting
duplicates because it eliminates changes in casing and separators
like dashes and spaces.

a side-effect of adding this fuzzifying function is that the ordering
changes slightly for some items, because while "Redis To Go" comes
after "RedisGreen" lexically; "redistogo" comes _after_ "redisgreen"..

the synchronizing is using our (now preferred) "interleaving" algorithm,
which requires a lexically ordered near and lexically ordered far
collection; and outputs a lexically ordered synchornized collection
as a result; all streaming. (the streaminess is the benefit we get from
the cost of all the ordering.). anyway, during the sync it will detect
that our collections are no longer in the "correct" ordered (with this
new keyer) and bork with a useful message. this commit both creates
and fixes that borking.

not a big deal but this makes me want to make the simplified key be
actual human key now, in its own column, separate from this 'name'
column; but this makes the markdown document more ugly and less
self-standing..