(-tag-lyfe.3.1 case study step 4: Synthesize roadmap)
the only structural change in this commit is that we added four
associations but note the impact it had on the overall structure
of the graph, visually.

we agonized a bit on this one: leading up to this point we were using
arrows (directed dependencies) in a way that was novel to us. rather
that the semantics being the familiar “depends on”, it now meant
“implies”. for reasons we haven’t yet wrapped our head around yet,
while these semantics are useful for states, they do not seem to be
so for actions.

another novel “innovation” (discovery?) was introduced here: rather
than having a graph consist only of action-states and dependency
associations; now we have actions only ever:
  - lead to states
  - depend on states
now the semantics of the arrow have reversed here: instead of an
arrow meaning “depend on” (as in “this action depends on this
precursor state”) the arrow means “is a necessary precursor to”
when used from a state to an action.

a side benefit of this is that arrows finally have the expected
meaning to the uninitiated (“this thing THEN this thing”).

but for whatever reason we end up with lots of little cycles, because
“state A” and “state B” have a reversed relationship directionally,
when compared to an action that moves things form “state A” to
“state B”.

so like “have mixed dough” is a state, “have baked bread” is a state,
“put bread in oven” is an action. “have baked bread” implied
“have mixed dough” (in the past). etc..

regardless, we finally have what we are after, which is something
suggesting a roadmap.