(-tag-lyfe.1.4 Create producer for hugo doc tree)
Discussion: the micro narrative over the last few days was like
  1. find any means of serving static pages via heroku
  2. this one heroku buildpack
  3. google “jekyll on heroku”
  4. google “jekyll vs”, discover hugo
  5. at some point discover github pages
  6. go deep into hugo but with a strong eye on jekyll and/or g.h pages

Not shown: In this commit (work) we discovered that yes, we can make
a symlink in the `content/` directory that points to a “real” markdown
file in our monorepo. what was next was to read the hugo documentation
to figure out if we can do something clever to inject front matter
into the “received” versions of our documents..

we’ll discuss the rest in the next commit.

(times)
  08-15 04:44  gonna start playing with top of the stack
        05:17  ok gonna try to index every hugo feature
        05:38  thought of multi-table and universal features
        05:53  decide yes you want the scraper script
        06:07  begin writing the producer script, realize
               you need to DRY up the producer CLI
        06:51  finished coding the abstraction from existing
        07:04  boilerplate code is done
               begin in earnest new producer script
        07:45  almost done with abstracting the soup function out. zz
        15:09  wake, resume
        16:51  ok we have a handle on the toplevel navigation element
        18:24  ok looks good enough maybe, the JSON dump
        19:44  now