(-tag-lyfe.1.1.11.2 Spike deep scrape for hugo theme TOML (*, 246))
Discussion:

Broadly what happens here is we sync meta-data from *all* the hugo
themes into our one “dim sum” document about hugo themes. as it would
work out, to do this to our discerning standards would require something
like 18 days of work (with some necessary long weekends rolled in).

To appreciate why we would endure such a quixotic detour, one must
appreciate how compelling the vision was:
  - we have ~252 alternatives (themes) to cull through
  - we have a ready-made system built around assisting in making
    informed decision,s provided that the collection of alternatives
    is tagged.
  - the collection of alternatives (themes) is _already_ tagged!

the vision of making possibly a wizard based on a flowchart-like
series of questions became an daydream obsession vision thing..

The thing was, we didn’t know the true nature of the metadata. and as
we dug progressively deeper layers into it we would find that in fact
it was less valuable at face value than we would have hoped. the
the three sub-component “reports” included here “prove” the various
short-comings of the metadata as-is. namely:

  - there are “features” and there are “tags” and there doesn’t
    appear to be a practical reason for them to exist separate
    from one another.

  - collectively these phenomena appear with inconsistent casing
    and styling

  - being that there is no governing body that dictates the
    authoritative set of allowed phenomena, alternatives are free
    to use willy-nilly tags and features. (experimentally we prune
    singleton phenomena before they enter our table.)

but we take steps to normalize and ameliorate the data here, bringing
us back to our (local) end-goal of populating our “dim sum” table with
clean, ostensibly high-value metadata about the themes.

On the Development:

We were too proud to break this down into further commit-steps but
maybe we should have. it’s kinda crazy.

One novel approach we took here: each ‘magnetic’ of our algorithm can
be invoked independently as its own command-line script. Similarly, our
“big report” has several “small reports” that inform it in its approach.

In all, there’s something like three magnetics and four “reports”,
totaling seven components each of which can be invoked on the command
line and each of each of whose output can be inspected individually or
piped elsewhere.

(times)

  09-05 18:46  a few minutes to plan (minus 3 minutes)
        18:55  upgraded pipenv and installed toml
        18:58  dinner
        19:21  resume
        19:31  succesfully parsing some tomls and not others
        20:09  ok we broke it down into 3 sub-magnetics
        20:54  about to start writing the first of three magnetics,
               change gears to set up its test
        21:31  ok back on asset
        21:42  really writing that mad parse
        22:25  almost done with one mag zz
        01:49  woke, resume parsing 3 individual lines
        02:27  first magnetic passing
        02:53  very happy with CLI output of magnetic
        02:54  and coffee break
        03:16  resume. start to play with integrating small scripts
        04:58  ok back from making coffee, about to do second magnet
               (the system call)
        05:32  being reading excellent khong article on the system module
        06:08  ok we have our first real output of the list of themes
        06:44  ok test looks good for second magnetic
        06:49  ok begin to write test for final magnet
        07:36  ok we are getting the one dictionary from the one guy
        08:22  worked around min version not version thing
        08:38  got a visual of every toml dump but one
        10:36  maybe an hour into the report
  09-10 07:50  after a few days of dancing etc, buckling down again
        08:25  have the basics of a subscription architecture laid out
        11:05  going well but zz
  09-11 10:03  begin
        11:20  zz
        16:26  rocking again
  09-12 02:54  ok finished the variations report to completion
        $(tm)  almost done with the report of uselessness. zz
  09-13 02:11  ok kind of tired but going to try to write the last
               report - most popular
        05:32  mostly happy with report - lots of little cleanup and
               we have to decide how to export the actual doo-hah from
               it, but zz
  09-19 01:30  been "depressed" for 6 days.
               got 3 tests working with ad-hoc memoizing class
        06:10  ok final report is good except isn't JSON
  09-20 01:58  ok sync worked - need to clean up lots of things
        08:04  now - WHEW