Explode mono-TODO list into many smaller graphs
the 100-line deep TODO.stack file had gotten ridiculous, as well as the
too-deep number identifiers on commit names.

knowing how far you are away from some end goal is nice, but it’s
coming at the expense of adaptivity. it’s simply not possible to be
able to know beforehand where your work will need to “swell out”.

as something of an experiment, we’re going to lean more heavily on
the graph viz files, which have the advantages:
  - human editable by hand with simple text
  - ride along with version control
  - natively model dependencies
  - more flexible in accommodating dependency changes
  - visualization

contrast this with our TODO.stack, which has some of the above
avantages but not others:
  - in its way it models a linear chain of dependencies as a list
    of text, but cannot do proper dependency graphs
  - not pretty for re-ordering

also (as mentioned) we went from our massive mono-file approach to
many smaller, more focused files. this will allow us to context-switch
as things come and go that we want to work on.

finally, the “ship of theseus” problem was what inspired this. OCD made
us want to keep that TODO.stack until it was pared down, so that we
could see those “git blame” dates on the lines when we finished them
(or do some other kind of visualization). we realized data is just
data.