Add single file schema (274)
Discussion: we always we knew we wanted this (different storage schemas,
perhaps this minimal case of a single file) but we didn’t think exactly
what it would mean for things like index files until we got down to it.
the below time entries form a bit of a narrative of our discovery.

We made a visual testing/dev tool that we ended up not needing for the
current task because once we made it we realized we don’t want single-
file schemas to have indexes. this works after a visual test but should
be covered if we keep it around..

(times)

  01:34  begin planning how to place tests for single file schema
  01:48  have first test whining about storage schema not found
         begin writing production code
  02:05  wow traverse was too easy. try create..
         abstract out “CUD” test support 
  02:14  amazing, abstracted support out of test file
  02:23  look in to making CLI that spits out an index
  04:08  realized we don't want an index file
         on a single file collection. 28 min later zz
  10:08  resume: injection for no index file (yikes)
  11:18  came up with schema “pather”
  14:35  finally have a CREATE in single file after soo many hoops
  15:08  have test written for delete
  16:41  delete ok
  16:57  started writing test for retrieve. 10 min later done.
  18:18  traverse. catch ninja scroll at michigan theatre a 9 lol
  20:44  resume. 10 min later might be all green
  21:29  now, begin committing. 7 min later done.