Add storage schema: two-deep (and 1st smoke test) (282)
Discussion: the bulk of the development work (that is covered) here
involves covering the index reading and writing functions to handle the
simpler case of when the number space (identifier schema) is only two
digits deep (so there is none of the stuff going on with having indents
anywhere).

What *actually* happened just now was that we gave the code our first
smoke test as a real CLI application writing and reading data. overall,
it succeeded with flying colors.

writing the necessary injections (for committing file rewrites and for
the random number generator) was exactly as challenging as we expected
and was mostly unremarkable.

the “three deep” filesystem schema “felt too” deep for just messing
around and the single-file schema “felt too” shallow so we decided
we wanted a *two* deep filesystem and identifier schema.

having covered two filesystem schemas already, we thought we would allow
ourselves to slide by on not covering this new storage schema with too
much attention; but what we realized was that the structure of the index
files would change somewhat too (exactly as we had designed them to
accommodate with the least change-impact).

so this ended up being the hardest part was simplifying the index file
writing function for this case (which *is* covered).

also we discovered a nasty bug: if the schema says it’s P deep but the
numbers in the index file are Q deep, the function would just decode the
numbers wrong in weird ways. now we assert the depth of each identifier…

what allowed us to debug this issue was the development CLI we add here.

one issue we did not foresee was a CREATE where one or more directories
need to be created…

another issue totally broadsided us: python CLI backslash-escapes
argv tokens WHY. the fix for this feels wacky, because we can’t easily
hit this in our tests.


MishMash:
- reallocate test numbers to satisfy OCD:
  - CLI starts at 800 now  
  - after the first few officious things, test
    numbering goes in triplets now (help, fail, ok)
- add handing for the “no directory” emission without cover for now
- have tons of fun writing the actual production function for committing
  a file write/rewrite. thank god it worked.

(time)
  04-18 00:51  first smoke tests
        02:14  after some messing around, things look good
        02:20  start playing with totally uncovered storage schema
        02:39  begin new storage schema with no cover
        03:19  bug in index, zz maybe an hour later
        08:11  after woke and coding for a little bit, after having
               discovered the issue in part with a dev tool
        10:52  with one case we are reading and writing shallow index files
        11:30  new schema looks like it's working
        14:55  track down this double-escaping issue. 15 min later zz
        21:16  from babysitting look at escaping issue and line up priorities
  04-19 23:01  family time was the wrong time to attempt that (that was 24
               hours ago. family time today too) but for now for real attempt
               to create a test to capture it
  04-20 00:26  the fix to the above must occur in production mode only. yuck