Learn gcc linking
CREATION MYTH: the FFVII trailer came out, and in response to it I
wanted to tweet out a hot take on it alongside a short video clip from
it. When I tried to use the software to download the video, it said it
wanted ffmpeg. When I tried to brew install that, it said it wanted to
get a more recent version of Xcode.

Once I had downloaded the hefty Xcode, I was reminded of my longstanding
dream to finally mess with swift. This pushed me to get “kiss-db” in a
good place so I could race into the next crazy step. I’ve been weirdly
leaving Xcode open because I’ve been unable to close it without finally
scratching this itch.

We did:
- at first we looked for tuts on swift
- that looked “easy” lol so we thought to learn about bridging
- this looks like it will be not so straightforward..
- took some tiny notes in our first ever “production” use of kiss-db
- figured out how to link C
- spike the typical project planning stuff including:
- some hacky but little hugo stuff

(times)

  05-13 03:41  contemplating feenix lol
        03:55  ok created our first note fragment
        04:37  start reading about app kit. zz while reading about it
        13:41  next day begin reading about integrating python w/ swift
  05-14 02:51  start a night shift, read re calling C-API's from swift
        03:10  read the short apple documentation on importing C funcs
               to swift. not enough info for a beginner lol
        06:17  AMAZING we learned something
        09:06  now