Introduce PythonKit
Discussion: we started searching with the term “interoperability” which
proved to be quite an essential keyword because it led us to all the
wonder and beauty that is PythonKit. It’s cliché but if it works from
the GUI it’s gonna be a “game changer”.

A working "PythonKit" means we don't have to write a C-bridge layer
which means a huge cost savings to our development.

Visual Tests:

    swift run --package-path pho CLI_with_PythonKit

    gcc -dynamiclib pho/pho-bridge.c -o z/libxxx-ohai-xxx.dylib
    swift run --package-path pho -Xlinker -L../z Pho

    gcc pho/main.c z/libxxx-ohai-xxx.dylib
    ./a.out

(times)

  05-20 04:39  begin reviewing our own notes to decide which tutorial
               to dive in to about making a minimal swift app
        05:01  our do-nothing app opens a blank window.this is thrilling
        06:36  great! did hello world GUI app and it went smoothly.
  05-21 03:33  having gotten PythonKit to work from CLI our mind is
               blown and we are terrified. begin.
        05:27  happy with our visual test. going to commit
        05:44  now