Introduce empty Xcode macOS Cocoa App
Justification: We are taking a snapshot of this empty app as-is (as is
generated by the IDE) so that we can takes diffs between it and our
near future work, so that our work is not otherwise munged in with the
generated code. This way we can, for example, try to reproduce our work
on some distant future iteration of the IDE/SDK and know exactly what
code *we* wrote that holds our “business logic” and design etc.

We added the "xcode project" directory to version control because A)
the Xcode IDE adds it to version control when applicable and B) we saw
someone else do it (the PythonKit package).

We added a path to a .gitignore that corresponds to a file in the
workspace that is binary data. It seems that we can live without
versioning this directory: we delete the directory, open the project
and run the the app and creates the directory anew. Presumably it
contain IDE GUI state information. We would rather not version this
directory/file because versioning binary data is gross and a smell.

We removed the cruft from two XCode-generated files because it's noisy
and the same info can be derived from our `#born` waypoint lines.