Update dependencies ((933))
While trying to establish that we had the latest toml we went ahead
and pulled down changes from our other dependencies (how?).

Something about something we upgraded caused TatSu to now trigger
warnings with the familiar thing about `collections.abc`. Unfortunately
the fix this came *the day after* the last release. (release v4.4.0
was April 23, on the 24th the fix went in to master).

Although we are able to install this non-released head of master
using pip alone, it appears we cannot use pipenv to get it into
our pipfile, etc (perhaps for reasons).

So unfortunately, to get this to work:
    pip install 'git+https://github.com/neogeny/TatSu.git'

which installed:
    commit 4da7dc98c7a1acb4cafd0e3f7c90c41e478ba877
    Author: apalala
    Date:   Sun Jun 2 10:07:18 2019 -0400

    [tokenizing+contexts] give tokenizer access to the ParseContext

Also:

  - beautiful soup stopped supporting ‘>’ the way we were using it
    in html selectors (because it uses soupsieve something something)
  - click doesn’t abuse exceptions as much, chasing our testing
  - toml changed one error message

(1+35+31+53+84+245+141+312+31)