Introduce toy storage adapter & canon (+13=325)
Discussion: If we can just complain for a moment; part of the whole
reason we went on a multi-day tangent around “tooling” (the graph
visualization, the well-covered single-document generation), was to
help debug a strange problem we were having with our content when it
turned into documents.

After several hours (reflected below) of narrowing the problem down in
our own code, we finally figured out that the problem was coming from
the vendor library. Several hours more of trial and error and we started
to reproduce the strange bugs that required what seemed like magical
incantations. It turns out the vendor library has some problems.

Because the state of toml support is not quite what we thought it was,
we started shopping around for alternatives. At various moments, we
considering settling for the too-heavy yaml or the not quite human-edit-
friendly json. we even considered violent xml or the old-timey unix
recutils. but them we found “eno” (to be referenced in docs soon).

In its way it may prove to be a happy accident because eno’s mission
statement sounds really similar to ours.

But we didn’t want it to have to be all-or-nothing. As quickly as we are
turning our back on toml we may turn around and flip back again. And
since birth, the project was never meant to be married to one storage
solution; we always wanted to do the storage adapter thing. So here goes.

But because this could easily become a hard-to-manage mess once we have
different storage adapters flying around, we came up with the idea of
a “canon” (commented in code here) that we usage uniformly (sort of)
against all the upcoming adapters. We wanted it to be sort of easy to
write new tests for “eno” (say) that leverage the work we’ve done for
toml.

But we felt that the best way to get this canon was to start with a
minimally simple blue-sky storage adapter whole only purpose is to
help drive the development of the canon and to act as a demonstration
of a compliant storage adapter developed with it.

(times)
  06-05 07:14  look for that bug in the docs again
               the bug did not go away
        08:11  we have narrowed the problem to one function
        09:32  absolutely a hack and needs improvement
        20:58  discover Reno
        21:25  gonna get to eno by way of a trivial yaml and adapter
               pattern or json.
        06:24  LET'S GO begin canon then use it for md table adapter
        07:17  wrote outline in comments for 16 cases
        09:16  begin md test cases in earnest
        09:45  begin implementing markdown table fellow
        10:35  passing one whole case! the collection not found case
        11:29  scanned lines and found no table
        12:57  proto parsing multiple markdown tables in a file
        13:41  good progress zz
  06-06 17:58  resume
        18:42  done with “resolve collection" 5 cases and 8 tests
        18:55  begin next test file: traverse IDs
        20:08  test cases are skeletoned out for traversal (two verbs)
        21:31  passing traverse IDs case
        21:41  begin asset code of entity traversal (making real
               entities). At some point got really wrapped up in updates
               and debugging new problems from that.
        06:11  now