Re-house “filter-by-tag” ((1029)) (437)
Discussion: “sakin-agac” was born with the objective of solving for
a static site generator (more or less); but it ended up as this
sprawling meditation on “data-sources”. Its most valuable by-product
became “tag-lyfe”, a tagging language and query language specification/
implementation.

“kiss-rdb” was conceived as a blue-sky project but had the benefit
of hindsight of the “sakin-agac” data sources, whose conventions were
cumbersome and whose scope wasn’t quite as ambitious as kiss-rdb’s.

THEN we made the “markdown table” storage adapter in “kiss-rdb”
nominally just as a prototype to drive the new API “canon”. However,
once we had *that* we found ourselves *really* wanting to close the
loop between “kiss-rdb” and “tag-lyfe”, so we could have the ability
to filter-by-tags on markdown tables, which would leave us with an
“app”-like ersatz version of an issue-tracker like we made ~elsewhere~
that accomplishes the most useful 80% of the behavior of that issue
tracker, but no new code necessary at all.

Specifically, we wanted the ability to take our issues listed in our
README.md files and just see (for example) the ones that are `#open`
or the ones that are `#wish` etc. It was too compelling not to
re-factor the integration of “tag-lyfe” away from “sakin-agac” and
on to the shiny new “kiss-rdb”.

For OCD completeness, we would refactor “sakin-agac” so it employs the
new “kiss-rdb” API (like possibly “sakin-agac” could be the owner of the
JSON storage adapter since its big thing was small standalone scripts
that output streams of JSON across processes using pipes of the
operating system). However, such an undertaking is decidedly *not*
compelling at this time.

Also:

* Add to the API (complexity): the entity class must now expose *two*
  methods/properties: one for the two-deep dictionary and one for the
  one. This is experimental, but the current rationale for this is that
  which one should derive from the other is storage-adapter-specific.

* Universalize “coll” not “col” as the short name for “collection”.

* The number of tests would be 1052 but we have actually *deleted*
  23 tests here! There were too many. Lots of asset code was painfully
  experimental too.

(times)

  07-03 08:51  begin work on re-house filter-by-tag
        11:40  over the hump
        12:07  passing one actual test
        13:06  we integrated toml but why
        15:24  passing visually, integrated w CLI
        16:03  done manually deleting old code, now pray tests ok
  07-04 12:20  (went really chill from above. travel day today) now.