5/14: No more `sys.path` hacking
Discussion: This one is perhaps the most embarrassing. At the time we
did it we knew it was wrong but we decided to borrow against the future,
and today we pay off the technical debt.

We wanted our modules (files) to require other modules (files) in a
way that wouldn’t require us to change the code that does the importing
when the client file moves; i.e., no relative imports, generally.

The problem became once we had CLI entrypoint files that were not at
the root depth. When these files were entered, the interpreter would
prepend the `dirname` of the entrypoint file to the beginning of the
`sys.path`. Reasonably the entrypoint file had no way of knowing what we
considered to be our “root” of things, so it always just prepends the
dirname of the entrypoint file to the sys path so that files and
directories sibling to the entrypoint file can be imported “for free”.

With `unittest` test runners we encountered problems of the same nature
but even more complex; probably we will spare elaboration here.

Our workaround solution to this at the time was to “normalize”
`sys.path` “by hand” so it looked the same way everywhere, regardless
of entrypoint level of depth. For one thing, this wasn’t even compatible
with the way the `unittest discover` works, and for another, it required
us to make an ugly amount of copy-pasted `[#019.whatever]` type files
that normalized `sys.path` “by hand”.

The right way to handle this issue seems to be to use pipenv, which
manages `sys.path` for you. Although it doesn’t behave 100% the same way
as our “by hand” technique does, it solves the original problem we were
attempting to solve, and we can finally get rid of all this noisy
boilerplate code.