11/14: Rename locations and identifiers ((-1))
(This is the least interesting stuff; the stuff that’s just purely
a by-product of all the name and identifier changes.)

One test lost:

We moved a test case from "sakin-agac" to "kiss-rdb" that was concerned
with the format adapters there. One test from that case was removed for
being no longer relevant because of a refactor and/or structural change.