13/14: Re-architect coverage tool for producer scripts
Discussion: ad-hoc change: don’t hard-code paths to things when you
can avoid it. Rather that the two-way association between asset and
test files; just use our now preferred [#010.7] case number convention.

(We put a finer point on it with this goofy coverage testing script.)

No need to track producer scripts with an identifier now that they are
all in a dedicated directory.