Introduce parser generator ((1035)) (+11)
Discussion: “cheap arg parse” is too weak. “click” is too heavy.
“tatsu” is wayyy too heavy for the scope of this… (The scope of this
will become evident in the next commit or two…) But mostly, this is
fun and thrilling to do the same old thing yet again but with a much
better sense of how to accomplish it with maximal power and minimal
complexity.

Also: move “throwing listener” to the home of listeners. 

(times)
  07-19 05:59  play with getting rid of schema row
        07:56  pause to go grocery shopping.
               rewriting "cheap arg parse" and its going well
        21:36  drafts for four cases of parser via grammar
  07-20 10:51  begin again with parser generator
        19:27  first grammar done
        19:58  other grammar just worked (3 cases)
        20:41  done with plural grammar
        22:12  now