2/4: No more sync-side entity-mapping: kiss-rdb (496)
👉 No more sync request context managers.👈 This was an over-complicated
artifact from when meta-data about how to turn the traversal stream into
a sync stream was emitting *as* the first record of the traversal stream.
The whole point of the current raft of changes is to change this API
point about syncing.

Terminology change: “stream request” not “traversal request”.
We now see it is a taxonomy where the general name is “stream” and there
are specific kinds of streams you may want: one for traversal, one for
syncing, maybe one for filtering, etc.

“producer script” sneaks in as an interface.

“markdown table scanner” becomes more formalized as part of our public
API for how files with markdown tables are parsed.

Melt EDIT's/todo's.