Rewrite pho CLI and introduce new eno roadmap ((1123))
Finally at the end of a months long tangent stack of unifications,
we took a big step back to revisit researching eno, our “great hope”
for pho. To this end, the main contribution of this commit is the big roadmap we added here. But also:

When we wrote the first pass of the pho CLI 5 months ago, we didn’t yet
have “cheap arg parse branch”, so we used click. Now that we do, we
prefer not to carry with us the cost of click here:

👉 Our syntax definitions are more concise and easier to read
👉 The “engine” is code we control whose API we can expand as needed.
👉 We suspect that our way is less heavyweight in a non-negligible way.
👉 We no longer have the many small annoyances of click here: extremely
   unfriendly for testing, for resulting in custom exit statuses,
   no help strings for arguments, weird things with argument names
   needing underscores not dashes ..

(times)
  11-10 09:05  begin revisiting eno
        10:03  done reading eno guide
        10:44  begin reading the python API documentation
        11:04  read beginning of python API documentation. seems good. 
               break. we might go top-down with document generation
        13:14  created 34-node roadmap lol
  11-11 13:32  re-wrote