Begin talking to google sheets ((1135))
Discussion: Came up with this name “transactors” as our take on
dependency injection. They can be plugged in to the collection so the
same interface is used (and same tests are run) whether the data is
coming from the cloud or from test fixtures. It’s the same
recording-and-playback pattern we’ve been doing for 15-ish years

Begin to see “primitive” instead of “string” in places because we
might experiment with storage-adapter-specific identifier API’s.
That is, we might want storage adapters to be able to work with the
kind of identifiers that are most natural for them when this step itself
is a “KISS” — namely, using integers to represent spreadsheet rows.

(times)

  07-03 16:01  begin google sheets storage adapter (end up refactoring)
  07-05 12:55  read in-memory stuff
        13:48  read a hand-written recording
        14:48  make recording
        15:01  read back recording of in memory
        16:32  snapshot of having recorded real live thing
        17:05  read back live recorded thing