Add google sheets insert rows transaction ((1139))
Discussion: we were suck on a bug for a long time not realizing that
we had to send date times as serial numbers (doubles). Had to read the
API documentation all over and just happened to find the spot where this
was documented.

Also:
  - correct frequent misspelling of “cell”
  - “subsheet” wasn’t a thing. we made it up.
    The terminology is: “spreadsheet” has many “sheets”

(times: this took at least 12 hours since last commit)