Introduce timestamp sync client working prototype ((+0=1153))
Overview:

Although not exhaustively covered & implemented, our “plumbing” API
seemed sufficient enough to make a CLI out of (without additional
coverage). Having exceeded our highest realistic hopes of seeing this
working well, we have accomplished the near-term goal. (Although the
client added here has no coverage of its own — it’s merely “glue” code).

Having said that, the page-long client code implementation feels a
little like a ball of mud, and we might eventually want to turn it into
a high-level function that can still somehow be tested.

Other Discussion:

The bash script gets a near-overhaul again. (Fortunately it wasn’t as
drastic as that in the end.) Adding the -sync sub-command (as an option)
and needing to pass through arguments *while* doing our other normal
option and argument parsing; it was feasible but keeping the code
clear & future-resilient has been a challenge. In this refactor we lean
heavily on the assumption that the empty string always evaluates to
false in conditional tests. This makes the code much less noisy.

Cheap Arg Parse is great but it needs more helpful error messages. We
need loosen up and maybe clarify implementation of our grammar grammar.

(times)
  14:29  begin work on live integration
  17:24  finished overhauling existing part of ts script
  19:33  after dinner and zumba, let's go, get to know exec
  22:01  was able to do exec successfully
  00:26  wrote really complicated integration
  06:47  live sync looks like it will work
  07:32  integrate with client
  09:55  begin see if the whole world is broken