( 3/16) Revert overzealous use of leading underscores
When we started using python we were in a "mood" where we felt we had
to advertise the privacy of the various names we used for all names
that weren't "API public". Likewise we had a .. somewhat peculiar
habit of lead-underscore-naming all of our variables that were
single-use. We have since decided it would be better to follow the more
universal convention of *not* doing this. (It's much prettier, for one
thing.)