(11/16) Overhaul syncing
Syncing is now implemented as an “injection” of a “flat map”. The
syncing agent creates the “flat map” and the storage adapter accepts
the flat map as an injection during its traversal. At each item, the
adapter passes the item identifier to the flat map, which sends back a
tuple of directives that do the inserts, deletes and rewrites given that
item.

Later we generalize this approach to work for all edits, not just syncs.

And:
- improved use of example row changes some test behavior
- remove anemic processor class thing