feat(pho:issues): the `top` command ((1270))
Discussion: the default report is top 3 open issues, in VCS
chronological order.

Side note: we became “self-supporting” in this commit 🤓. Issues we
opened up in this commit we did so through the CLI.

Wish: currently “group by file” always takes precedence over sort
by “mtime” (for practical reasons). It would be interesting to be
able to see the oldest open tags in the universe etc. But it would
be a little involved. Table rendering is set up for this, though.
Crazily it could be determined by the order the opts appear in.

(times: 12:45-15:48)