refactor: synthesize scanner lib from the universe ((1306))
Discussion: We finally snapped when re-writing yet another
“scanner via iterator”. We unified around the name `next` (instead
of `shift`) because that’s the verb that java scanner libraries seem
to gravitate towards. We like the `more`/`empty` duality even more now.

Usually when we turn something into a low-level universal library
module like this, we want to piggy-back on some very old module
somewhere to “keep DNA”; but in this case that felt like “string
scanner via…” deserved to be left alone and our OCD warning light went
off so we just bit the bullet and birthed a brand new module for this.

(times)
  10-08 19:57  start crazy town
  10-09 05:53  crazy town seems ok enough to start using
        06:54  going well but need nap
        10:27  resume first test
        11:55  begin lines for CSV
        11:12  begin tangent: unify scanning
        14:23  done
.