refactor: no more RECEIVE_DIFF_LINES hack ((1375))
Discussion: When we originally constructed our API for the result
signatures of our three canonic edit functions (update, create, delete),
we did so with a KISS-ism that now, with the benefit of hindsight, seems
to have been a bit overzealous:

the three canonic edit functions (update, create, delete) produced
results consisting directly of the entities they involved:

For CREATE, the result was the created entity; for DELETE, the result
was the deleted entity; and for UPDATE the result was at two-tuple
of the “before” and “after” entity.

This convention met our requirements at the time and hummed along nicely
for a few years. But trouble brewed when we started trying to simplify
and unify things around single-file collections: It became more and more
convenient to want there to be a *returned* patch file representing the
change to make.

Up until now, we had a hackish work-around as a way of getting the diff
lines back from the storage adapter: they called a callback that was
attached to the `opn` function (something that itself is on the chopping
block for being too hacky).

Our newest solution is that the result value from these three functions
is a *custom structure* that has fields for the participating entities
AND things like the diff lines (if it’s a single-file format).

There are by-products of this change that we expected to see more of in
this commit that we aren’t seeing. (Where do patches get applied in
production? Do they?)

There is also a tension that is introduced: we want the diff lines to
be accessible for testing, but “in production” we want them to
“just work”. In a subsequent commit we’ll try to confront this tension
with approaches like using the “shell/kernel” architecture some of which
still remains in our SA architecture…

Other disjoint comments/observations:
- Our only real life workhorse directory-based storage adapter (eno)
  is insulated from a lot of this change because it doesn’t even
  implement two of the three canonic edit functions - it gets it work
  done mostly thru the more sophisticated and modern “batch update”
  function…
- An asymmetry is discovered where some of the CUD implementations
  bundle their own “emit edited” emitter and others don’t

(times)
  11-11 09:20  refactor for the patch thing
        11:25  zz
        13:00  wake & resume
        15:38  break for run then zumba
  11-12 03:00  try again: what am i doing
        06:28  all green!?. zz shortly after
        12:22  now
.