docs: absorb “storage roadmap” into the “everything” table
Discussion: This is a pretty major milestone before a more major
milestone: this is absorbing one graph into another using our new way.
The next step will be absorbing a physical whiteboard into the table.

This reveals one new issue but we’re going to go ahead anyway.
It appears that a node will be rendered inside a subgraph if it
is first referenced inside that subgraph (makes sense). This is causing
some undesired rendering because of whatever rationale we have for
ordering subgraphs. What we would rather have is render all deeper
subgraphs after all shallower subgraphs and the outermost layer.
But later for that!
.