docs: absorb physical whiteboard into the “everything” table
fix(pho issues graph): render assocs after all nodes (34)

As for the fix, if we don’t render all assocs after all nodes,
then the first occurrence of the node *reference* determines its
placement, in terms of the subgraphs. We want the node definitions to
be the first time the nodes are mentioned. To ensure this, put all
references after all node definitions. Otherwise, nodes end up in
subgraphs they don’t belong in.

As for “absorbing” our physical whiteboard, the main goal here was just
to get every high-level project idea we’ve ever had over the last few
years into the same file. The structure is scattershot and there is
some redundancy of nodes; but nonetheless it’s good to be able to see
everything in one big document for once.

Amusing anecdote: we did something crazy in vi that made it appear
to reverse or rot13 or our code lines (just on a few lines) so we
threw out our work and found a much simpler solution than what
we were about to do.

(times)

  22:35  begin absorb physical whiteboard
  22:52  33 rows data-entered (no assocs or tags yet)
  23:33  done data entering and assocs and tags (very rough draft).
         begin rendering bufix
  23:59  amusing anecodote. we might be done fixing bug.
.