test(curses-yikes): introduce beginnings of simple components (23)
Discussion: two-pass rendering to terminals is something we’re familiar
with from rendering tables and table-like data (most often when we
line-up something like an “options” screen in a help screen). This
shows the beginning of that in a highly abstracted way.

This is very much a WIP snapshot. For whatever reason we are eager to
jump ahead to the more daunting part of all this: our state machine
stuff…

(times)
  23:25  begin simple components - eventually get to state machine
  06:10  looking good. rendering blocks of letters.
         need proper rendering and states
  11:46  slept from previous time to about 10:30. close off WIP
.