test(curses-yikes): visual test: flash and introduce static buttons (36)
Discussion: It’s thrilling to finally see components rendered to screen
with some dynamism. Not having the input loop covered feels bad. Will
probably push things around to address this soon.

Because we got this far developing all these components *without*
coupling to curses *at all*, now that the time has come to couple it,
we don’t want to build the dependency too deeply in to it… This kind
of thing might lead us to keep all curses-specific stuff in one module.

(Currently there is one curses-specific facility at the topmost module.
We want to move it to the the “curses adapter” module. but this would
leave the topmost module almost empty lol.)

(times)
  20:52  begin visual test: buttons and flash only
  22:02  ok it was as we feared, debugging is hard. we got our first 
  23:56  finished tedious rewrite of nav bar render
  00:11  finally rendering nav and some custom buttons and layout
         looks good. now, the loop
  01:23  got flash to write a real message. we love it
  01:29  we think we are done with visual test even though no static API
.