test(curses-yikes): introduce orderable list (60)
Discussion: we anticipated this step as being the most difficult and
it has proven to be so. But nonetheless, the control is now existent
and 85% okay.

State machines seem to demonstrate their efficacy here but, there are
bugs in the control (not covered) that exist intertwined with FSA-ism.

We’re holding off on our final judgement of them until we get through
injection.

After a couple days of re-architecting things to accommodate the
“controller stack” and figuring out which parts needed to become
dynamic; we went thru several phases of re-architecting in our tests:

1. The two-stroke (multi-stroke) processing of input makes tests
   overly verbose and hard to read (and probably brittle too). We want
   to be able to test at a higher level more like (yikes) cucumber.
   So we began with creating this “long story” facility. But still,
   the main glaring problem existed that our tests stories were
   overlong for one test method.

2. The solution to the above ended up pushing the application arch.
   forward a bit: We want to be able to (in effect) “serialize” the
   relevant state of the interface and “unserialize” it. This meant
   representing and re-constituting *focus* and also the equivalent
   of form data (for multi-dimensional structures like lists). This
   allowed us to break a long story up into “waypoints” and at each
   waypoint we do some interaction that moves the story forward a
   little and then make assertions against the screen output. 


Also:

After a lot of consideration, we re-arranged our lexicon a bit: when
we key-up and key-down to change what component is being pointed to,
we do not “select” it, we give it “focus”. We decided that “selection”
should be more of an opt-in action from a user than simply mousing-over
(or equivalent) (and indeed there may end up being no employment of that
verb at all here). 

We were hesitant to call our concept “focus” because that term felt like
one with such a strong pre-existing meaning for a true GUI which we are
not; but the more we thought about it, the more we realized that it’s
a close-enough fit.

(We went back in time and renamed “selection controller” to “focus
controller”. Traces of the old verb-stem still remain in some places.)


MISHMASH:
- don’t use an attribute called `state` for anything, publicly
- when we use to say `is_interactable`, we may have meant
  `is_focusable`. The old term remains in some places, as its more
  general and still makes sense semantically.

(times)
  12-02 02:34  begin writing test file for orderable list without having
               a strong sense of a spec except in head
        03:59  awesome but zz
        09:35  resume, implement stack push
        00:05  gonna start the visual test now. we gotta.
               we need that emacs mode to be integrated to our fields.
  [12-03 - 12-07] [tons of work. 40% on arch, 60% on testing arch]
  12-08 08:38  now
.