refactor(curses-yikes): get FFSA’s out of button processing (60)
Discussion: Earlier iterations served as a proof-of-concept that we
can derive interfaces (partially) from formal state machines. Now that
we have “proven” this, we refactor it so that this becomes merely an
implementation choice, rather than being baked deeply into the package.

Specifically this affects how we survey components for their available
buttons associated with their various states. Now the component AAs must
expose an API method. This will be more necessary when we deal with
injecting item classes, which will be complicated around having a
“one state machine to rule them all”.

(time)
  12-08 20:24  begin analysis: traverse FSA's thru a method, not directly
  12-10 13:27  passed all tests
.