step(pho): snapshot of minimal, 2-page pelican project
This has no special theme, nor does it pull in all the files generated
by the pelican-quickstart. It’s just enough to generate the site
without warnings.

For some reason it’s generating both pages even when we only request
one with `--write-selected`, but we’ll look at that later.

The main purpose of this is to have a pristine, known working state
as a waypoint to com back to as we develop, sort of like a visual
test.

This is a development tool and .. probably will not be useful to
the broader pipeline, but then again maybe it will when we auto-generate
intermediate representations!

But for now it’s just a “save point” as we delve into the fun of
adapting the Alabster theme…
.