feat(pho): TMX Notebook enters bleeding alpha ((1483))
Add to “requirements.txt” livereload and tornado so we can have a simple
static web server that does livereload (not depending on pelican).

This commit realizes most (if not all) of the architectural ideas
sketched out in [#409.4]:

- A custom tcp/ip service that listens locally for ad-hoc business
  events. It can cache stateful information about things, like
  parsed config or it can deep track of times for debouncing
- a watcher (for now using `watchexec` which we mostly like) that
  watches any arbitrary directory of our markdown files in our
  sort of custom format. The watcher `exec`s  to a new process that
  (as if invoked by the command line) connects to our tcp/ip service
  and tells it which file changed
- The generation service does two hops: one it writes to an
  intermediate SSG project in markdown compatible with the
  SSG (pelican). The second hop is that it generates the final
  html output using the SSG. (Currently it only generates one file
  at a time and when we want to generate the whole site we have to
  do it “by hand”. This will probably meet our 80% use case (MVP lol)).
- Here we also have our own simple static web server that listens to
  the above html output directory and does live reload on it.

The end effect is that we can be editing one of our TMX-flavored
markdown files, press “save”, and we can see the web page reloaded
live reflecting our changes.

This effect isn’t that dramatic now, but it will be more dramatic
(and useful) when we get it working for TMX notecards…


Other Changes

Our internal API for config commands allowed them to yield output
lines, which was nice and convenient at first but proved to b
limiting when it came time to fail gracefully. Now, errput output
infoput etc all go into the listener and the result is always a
returncode integer.

(we lost our editing buffer because ‘W’ is right above ’S’ and we
mistyped it (thanks GitX 😏) so we lost our times, but we’ve been
working pretty steadily on this since the last commit. we really
wanted to reach this point before the ball dropped on new years,
but two days behind isn’t that bad.)
.