test(pho): derive document via declared demarcation ((1487))
Discussion: we began by imagining “what if we were watching a directory
with eno documents (a kiss rdb collection) rather than a directory of
TMX-flavored markdown documents?”  As a rough prototype we added this
“extension” option to the watcher, and injected this extra parameter
throughout the stack.

(This is exactly what we meant by “case matching” in our one big
architectural document: take the file that changed and look at things
like its extension to decide what course of action to take.)

Suddenly we were struck with: “since writes to notecards happen
‘through our database’ are we being crazy to watch the directory
instead?” but our fears were quickly quelled because we remembered
that a target use-case is editing eno files directly.

Thinking forwards from the above, we imagined that probably we will
want to get from “this eno file changed” to “what entit{y|ies}
changed in that file?”. There are approaches to this ranging from
lazy to byzantine. Lazy: just assume every entity in the file changed.
Byzantine: keep a lines cache in memory (or on disk) of EVERY eno file
at server start, and when we are notified of a file change, do the
diff ourselves. Our in-between solution is to use our VCS to get an
answer; an approach that will have PRO’s and CON’s we discuss in the
next commit, probably.

Leaving the above as a black box for now, we induced that what we want
is to be able to get from any arbitrary notecard to a an “abstract
document”. Now, “abstract document” is a new conception we came up with
only in the last few weeks. The 17 months old legacy code was geared
around turning a TMX-flavored markdown document (derived from eno kiss)
to writing a hugo-flavored markdown file on disk. Straddling the old
way and the new way will be annoying.

(times)
  01-02 23:24  started writing directory watcher for eno trees
        01:43  begin trying to document the model
  01-03 12:00  fixture done, write tests for new way in new test file
        13:51  finished big step one of two of the algorithm: we found
               the document root node by traversing upwards, now we have
               to inflate the whole tree from the root node
        14:57  two test cases green for declaration-based demarcation
        15:42  now
.