test(pho): introduce adapter-aware CLI for new generate ((1529))
Discussion: We spent a full day just prototyping around and trying
different stuff for our new "generate" interface that was:

- now adapter aware and
- unified in interface for:
  "website", "markdown", "dotfile", "tree", "check"

We ended up with lots and lots of code and ideas.

As is so often the case, breaking it up into reasonable sized
(and testable, and non-breaking) commits is the hard part.

This adapter patten shown here with dynamic CLI's based on plugins
activated in the selfsame CLI, this is a thing we do over and over
again. Implementations of it exist in antiquity, just FYI.
.