test(pho): new big index ((1541))
Discussion: Before the massive disintegration of our big ball of work
into separate patch/steps, we had a full new rewrite of “big index”
that took in to account the hierarchical container type… But some time
between finishing that and a few days ago, we realized that SSGs would
need more from the big index to decide whether they can generate
a markdown from them; and also something important (order) was missing
by merely determining what the documents were in the collection.

During a run in fact, we developed this idea of the “document depth
range”. If we were targeting sphinx, or some big book or something,
we could have documents at different *depths*: “chapters” (say) might
not just all be a flat list of documents: they might be broken up into
sections and subsections and so on. (Also, we probably wouldn’t have
the idea of “documents” so much as sections with automatic pagination, 
but ignore that for now.)

Axiomatically, we say that if you’re holding a node and that node *is*
a document; then it *has* a document depth of zero. (The number acts
sort of like an offset in an array: If i’m *at* the current document,
how much depth do I have to add to get down to a document? In the case
when you’re already *at* a document, the amount is zero.)

So if you were holding a “chapter” node, and that chapter was just a
flat list of documents, the document depth would be *one*.

If the chapter was broken up in to sections (but not subsections); the
document depth would be two, and so on, but we don’t just stop there:
It’s entirely common for some sections to have sub-sections (and maybe
sub-sub sections and so on) but not others. This is where we derive
the idea of a document depth *range*. (We call it a “minmax” in the
code, out of deference to RPG videogames.)

We’re going to dump more documentation sketches her that we decided
not to fold into the main body documentation. But before we do,
we want to mention that we spent a whole day just on writing the
pseudocode for this, this one local maxima of a mountain, then a full
few hours writing the new asset (a second rewrite (the first never to
be committed)), and a full day on tests, and one test in particular
realized a gaping hole in our algorithm (explained in a case) which
we then covered in asset.

As for the data model and the big index narrative:

We *think* what happened was, in the original data model, the only thing
we knew we wanted for sure at first was that notecards had a `prev` and
a `next` (to accord with the rough GUI sketch in the README).

Although internally we could have modeled this as only a singly-linked
list and derived the other relationship from the one (e.g. determine
every "next" by seeing every "previous"), this would have required
either some on-disk derived index or taking a big index of the collection
on every read of one entity; both of which are not KISS-compatible.

As such, every entity has both a `prev` and `next` field.

One the GUI matured, the time came to allow entities to have parent
child relationships. At first, it seems, there was no `children` field,
just `parent`, because somewhere at one point there was this idea
that children did not have a sense of order intrinsic in the model.

But the main point is, children *do* have order in the datastore,
but still we have a `prev`/`next` coexisting with parent/child in
a way we never really understood until this last swath of work:
The short of it is, the two associations are redundant but we take
steps to enforce that you only use one or the other or a particular
combination of both.

In the New Way, big-indexing is still mainly to answer the question
"what are all the documents in this collection?" but it has gotten more
powerful and complicated: documents now get their order and structural
relationship from the notecard relationships just like was always
intended. Documents can be deeply nested within what we call a
"node tree". (We may call it a "document tree" if it has documents in
it.) Probably the only rule here (beyond basic data integrity stuff
around ordered trees, no cycles and so on) IS, you can't put a document
inside another document. Otherwise you have freedom (for now) to
nest documents arbitrarily deeply within “document nodes” and then 
(for now, and/or sort of) you can nest nodes arbitrarily deeply within
documents. (There *is* a hard-coded limit of 6 somewhere for good
reason, but in theory we could go deeper.)

So for example you may group your documents in to chapters but also
break each chapter in to sections and so on. You may decide you want
sub-sections and so on. There are no rules here with how deep your
hierarchy goes, or whether the depth needs to be consistent across
the collection or otherwise.

Finally, it’s the SSG adapter that gets to decide whether they can
handle any given document tree with some given document depth range…
	
(times)
  01-24 09:05  begin think about requirements for new big index
               and what test cases
        11:47  done editing and rewriting and re-housing documentation.
               begin to write one massive god case
        12:06  finished very rough sketch of all our tests in 1 big case
        13:53  we have the old new index working on the big crazy new
               fake graph and it's amazing. now we have to make the new
               new big index
  01-25 21:27  making the big index of the single node tree was so
               satisfying we flung off in to space and went running
        21:45  start big index of whole collection
        23:29  ok wow the big index of the whole collection passed
               rough draft except we aren't exercising built hash
  01-26 22:33  begin assemble commit
.