refactor(pho): add “check” command ((1541))
This is mainly a flushing of a bunch of small stashes.

We gotta call this a refactor and not a “feat” because we don’t actually
cover the function itself for now, it’s just “tooling” (but note
inline code for visual testing lol).

We set out to update our old visualization (GraphViz) to work with the
new datamodel (and actually render deep document trees); and we set out
to fold in a new visualization, the text-based one that looks like
`tree`. But just below this stack of two visualization commands is
"check", the one that simply creates the big index. Just folding that
in (and folding-in misc stashes) was enough work for one commit.

Note this is called "check", but what happens when the collection
fails integrity check is just a bunch of hole with one-line "write-me"
type messages (that are anyway somewhat semantic).

So:

- cleanup, modernification (“open” not “to stream” even when you know..)
- one bugfix (that change from 0 to -1 somewhere)
- begin to integrate new big index into mainline (it’s too convenient
  not to make an accessor method just to avoid the coupling)
- starting to avoid the bare name “collection”, also the name
  “notecards”, when internally “business collection” is the only thing
  that’s immediately unambiguous at a glance
- start to distinguish explicitly whether our business collection
  is mutable or not, from the beginning, since non-mutable (caching)
  collections will be more and more of a thing

(times)
  01-27 00:56  gonna start on the two graph commands
        01:58  excited to do 3 things but take a rest
  01-28 14:17  finished "check" after about 3 hours today
.