feat(pho): add “body function” field ((1563))
Discussion: this was a little harrowing for a bit: found that we had
been chugging along without parsing schema files on directory-based
collections for quite a while. Added a special option to tell kiss
explicitly not to look for a file, to maintain some legacy tests
ick/meh.

Thoroughly enjoyed writing a parser for our legacy markdown files
yet again (really!)

  02-04 11:00  begin specing body function
        11:43  begin reviewing .rec format again
        13:39  notecard class makes first contact with
               new field called “body_function"
        15:31  begin implementing fall custom function, in earnest
        22:45  after a run and several hours of reworking kiss to parse
               schema file even when etc, back on track
        02:50  great progress but gotta zz
  02-05 09:56  resume
        15:12  begin whole world integration; making commit
        16:38  yikes finished after lowlevel rockiness
.