refactor(kiss-rb): rewrite md table parser to use FSA ((1567))
Discussion: there’s a feature we want to add to pho “issues” and it
requires that we access the “interstitial lines” between the header
line and the markdown table, of the issues. The existing code was a
huge fragile mish-mash of scanner-based code, so we rewrote it using
our favorite pattern for such things, this one FSA pattern.

We haven’t added the feature yet: this is just a case-study snapshot
to show the before and after…
.