feat(pho): make “issues” fun to use ((1569))
Discussion: the “issues” utility hobbled along without getting a lot
of use for years because it was a bit cumbersome, and because it didn’t
simulate what we did “by hand”: allocating the right identifier based on
several criteria. We fixed the allocation issue a few commits ago, but
still it was cumbersome to have to specify the path to the readme file
or set an environment variable. So now we allow persisting to the
filesystem, which makes using this actually fun now

Also, fix that help screen issue.