content(kiss-rdb): fold kiss-rdb docs into mono-site ((1569))
Discussion: We didn’t know which way we wanted to go with this at the
outside: should “kiss-rdb” docs be targeting a separate site than those
from “pho”? In a first pass, we proceed as if we were ~ keeping it
simple ~ and targeted making a separate site. Then we signed off that
the generated content looked good.

Then we attempted to merge the trees which is where it got interesting.
Our temporary solution (which we are starting to see as a final
solution) is that we created a new body function that pulls in remote
content from remote notecards vaguely similar to how content is pulled
from files on the filesystem.

New features: more helpful error messages when we run in to RSL issues.
Before we had no context for where the error occurred. Now we can know
the NCID.

Performance Discussion: as anticipated, pelican is starting to show
a noticeable performance hit as the number of documents grows (now 60).

“pho” still produces the content documents “instantly” 😏

(times)
  02-09 18:41  ok i guess we'll look at structuring the kiss rdb docs
        20:32  finished entering 25 paths by hand
        21:09  now we have the 15 markdown docs in a flat tree
  02-10 12:56  now