content: add all remaining but /doc ((1569))
Discussion: earlier we weren’t sure but now we’re sure we want to see
all of the DTF TMX library in one place. In this lap we conceived of
doing the “expanded by two levels” left nav.

Also we started categorizing (softly) our nodes with names like
“Library: “ etc in the title.

Saving this now even though there’s only a little further to go
to get all the docs so we have a regression point in case we choke
on memory.

(times)
  13:10  start to count up remaining docs
  15:01  done hand-entering 13 more docs into graph, w/o content cleanup
  15:10  tree parses, now we have to structure it correctly
  20:57  now