feat(tag-lyfe): simplified tag query parser ((1575))
Problem: tag query parser written with a parser generator is heavy
and obtuse and loads a large library that causes CLI clients noticeable
latency. Furthermore its an external vendor dependency we’d rather not
rely on if not necessary.

Solution: For our current expected use-cases, we are happy to use
a much smaller featureset of a query grammar: we just want AND and OR
and parenthesis and just basic tags. This grammar is small enough to
parse reasonably with one of our hand-written FSAs following its
particularly popular fashion

Also:
- this “wordables” thing felt obtuse. was just “to_words” but backwards
- the ASCII tree thing is so small, we would rather not abstract it

(times)
  02-12 00:06  begin adding priority
  02-13 12:58  done making prototype for new token query parse,
               gonna split the pile now
        15:16  ok lets commit
.