dep: `yarn install`
We’re following an explanation in documentation linked to in our
notecards that says we don’t need to commit the (10k SLOC) .pnp.js
file if we’re willing to run `yarn install` when we switch branches.

Since we’re single-developer (and frequently single branch), we’ll
abstain from versioning this large file for now.

.