wip: achieve React/Rollup “hello world” ((1614))
(This is all thanks to the excellent blog article by Nathan Sebhastian
as cited in documentation added in this commit.)

Here we introduce an ersatz “namespace” in our top-level “package.json”
scripts. Yes this is a smell. But we maintain this single top-level
package.json file both for the convenience of (mostly) always working
from the same directory, and to keep all our npm-reliant sub-projects
using mostly the same toolchain and same versions, as sort of a
poka-yoke to discourage erosion in those projects.

- Absorbed roadmap-style dotfile into issues in README.md
- “pho issues” practical hotfix: ignore ‘#priority:M.N’ when making
  graphs. (This still has a non-critical bug where it lets the number
  through.)
- “pho issues” hotfix: how was `--format json` not covered? 🤔

(times)
  09:52  begin studying react formally
  14:28  hello world!
.