wip: achieve tic-tac-toe (step 1) using Rollup
This is a "waypoint" (a "save").

We are about to switch to *not* using Rollup (for now) so that we
follow the "orthodoxy" of the "official" tutorial from the React
documentation (that we must treat as orthodoxy while we don't know
what we're doing).

But wanted to save this state for posterity for if/when we change
back to rollup, so we can get an idea for how it was working.