refactor(kiss-rdb:md): no more Action Stacks ((1615))
In order to make a small “todo”-style graph (like our `.stack` but
with dependencies), we wanted the ability to “comment out” entire
markdown table of tasks (and certainly not in a way that requires
touching every line.)

We knew we had hacked a little bit with table “directives” somewhere
so we thought “it should be easy.” But as it turned out, it was a costly
feature to add along several vectors:

confounding vector: the “action stack” API (although not without its
use and novelty) is fundamentally incompatible with multi-table files.

confounding vector: turning iterator into scanners (a day of work) ended
up breaking deeply the “context stack” facility (today’s day of work)
which only works if there’s no lookahead (and a workaround/fix doesn’t
easily present itself).

Although we want to improve this work further, we are leaving it as-is
in its non-refactored state, to justify the API we are about to create.

(times)
  05-24 23:50  begin add support for multiple tables
  05-25 12:52  next day same thing
  05-27 04:10  now
.