rename: Rename .stack -> .ruby-stack
Make room for the other file of the same name from TMX-python.

The last edit to the former file was late December, 2017. It is
now January, 2022.

We considered resolving the merge conflict by stacking all the
new items (from TMX-python) on top of the old items (from
TMX-ruby) but, we didn't entertain that thought for very long:

Most of TMX-ruby is dead weight/archival/historical at this point,
and probably we will do something like sunset most of it or prune
it out somehow, or move it to an issues file, etc.

.