refactor(pho): Clean up identifier allocation ((1619))
The code for this was unworkable and had bigs. We have rewritten
it for clarity and uncovered those bugs.

This was a work in progress from sometime in late 2021 that we put
on hold for a few months. Its original intention was lost. Probably
we were trying to make it usable for opening new issues.