feat(kiss:cap-server): Introduce redbean (lua)+python proof of concept ((1628))
Summary: Formalize what we witnessed in our unversioned one-off from about
a week ago, that we can in fact "call back" to python from the lua in
redbean to make a web application.

Also: Introduce two recfiles that we intend to use later as data files
for this "capabilities server" project, which (in one incarnation) we
would call something like the "polygot programmer's cheatsheet" phrasebook/
dictionary. (Wow we need to find a name for that.)

Disucssion: there were two instances where we got really stuck and then
finally unstuck. Each of these instances consumed the better part of a day,
(with hacking time interspersed by things like restaurant work).

We want to take a second and write about it to make a
little case study of how this happens and how we got unstuck.

First, we were having issues with passing multiple arguments in a call
in bash. Second, we had an issue doing the same sort of thing in lua.

The issue in bash was from the line (recommended by the lua docs)
of using `bash -c '[etc]` as a  "workaround" for invoking the redbean
server from zsh. as we write this we realize we didn't need to use this
form (probably) because we were not "in" zsh but in a bash script.

We needed to figure out how to effectively `some_array.join(" ")`.
It was by doing something like `"${some_array[*]}"` which was confusing
and hard to search and hard to debug and understand (as is characteristic
of bash programming).

On the lua side, in order to figure out how to handle variable-length
formal positional parameters to a function, we googled something, it
wasn't working, eventually we figured out that what we googled was for
an earlier version of lua than our version. This served as food for thought
for how we would wan't our "polyglot hacker phrasebook" project.

(times)
  03-09 17:19  begin start server script. 5 min later nap lol
        20:39  after nap, script is on its way
  03-10 08:49  quick few minutes from work - parsed argv, now do the
               interesting part
        18:41  wow used exec to start server process
        22:36  got start server working with difficult bash stuff.
               now needs install advice and find minimalest server
  03-11 08:54  (for one hour from GC) found target redbean build,
               wrote installer explanation script
        14:03  from SW Plymouth, start writing "index.lua"
        14:34  calling our python ping from the lua webserver
        14:56  it went well. about to make function. populating cap server
  03-12 21:15  from Amanda's, write while loop
        22:13  wow finally got the page to render thru lua