rename: Move one file from pho to kiss
Discussion: When we initially created this recfile, we put it alonside
the one about learning lua; both of which we housed in "pho" because
we envisioned it being a blog-like "article" (adjacent).

In fact, the "capabilities database" about recutils is pretty close to
the "kiss-rdb" project (more than "pho"), because the database will help
guide the development of the recutils adapation rather directly.