feat (kiss:cap-server): Make formal schema via CLI lol (+0=595)
Rough sketch. Pure insanity but we couldn't resist the urge:
Make a CLI for parsing an abstract schema. The CLI outputs S-expressions.

Make functions for consuming the S-expressions. In this manner, we could
write other modules/functions like "html form via abstract schema (entity)"
and we could fine-tune, trial/error the form generation thru the CLI,
by piping lines of S-expressions out from one script and in to the other.

This work here is sadly not covered (YET) but here's the visual test:

    py kiss_rdb/magnetics_/abstract_schema_via_sexp.py \
      -formal-entity AA -attr BB -key -attr CC -optional \
      | py kiss_rdb/magnetics_/abstract_schema_via_sexp.py -file -

The first process parses the CLI and outputs a sexp. The second process
consumes the sexp in STDIN and outputs the same sexp. (This behavior
is not useful here except for testing that it's lossless encoding/decoding.)

This will be more interesting when we make other processes that consume
this sexp input and output something else, for example HTML or SQL or
the begining of a recfile..

(times)
  03-18 12:49  (from SW Plymouth) begin read about joins, review existing
               formal entity structures
        13:18  we have some sense of how we will use recinf and typedefs.
               begin reading about joins
        13:54  Join wouldn't work for our example but we have it working
               for the example in the doc. narrow it down with binary search
               leaving SW Plymouth because tired
        14:14  (from home) figured it out - you use the local not remote name
  03-19 08:20  (from work) make each formal attribute as you go
  03-20 08:53  (from work) rough version smoke test of generate sexp
               (pretty print) looks great
        20:37  (from SW Plymouth) begin read sexp
        20:52  began to write read sexp lines
  03-21 18:27  (from home) finished most of sexp consuming sketch
        19:19  (now) smoke test rough sketch works