feature(script-lib): Engine enters feature-complete "alpha" +0=23
We didn't like how we had to allow "-" though as a non-option-looking
token at a very early stage in the pipeline to allow the participating
formal node to consume it as an argument value. So (experimentally)
we changed the whole "tokenizer" to recognize the special tokens of
"-" and "--" which are not considered proper options NOR proper non-options.
now; but rather "special" tokens. The formal must now be defined with
a special meta-flag turned on to allow it to consume the "-" as a value.

This completes our first-pass run at the "engine"; now we get to make
frontends and the help screen.

(times)
  04-04 18:20  begin bang the hell out of -file
  04-05 00:01  happy with it
               mosty. .. considering changing tokeninzing
        01:25  wow done
        01:58  now