feat(script-lib): Spike engine "help", 2 integrations +8=44 (293) ((1683))
Discussion: integrations have begun because we needed something real to
do visual testing with so we chose two scripts with some of the simplest
interfaces from our punchlist.

Getting "docstring 2 description-lines" was remarkably byzantine

There's a bit of messiness that creeps in here as we attempt to stitch
together our backend and our frontend but it's acceptable

Added 2 test cases for oversights discovered when visual testing.

(times)

  04-14 22:58  (from home) begin integ 1/N
  04-15 18:18  (from LAF parking lot lol) integrating
        18:22  (done)
  04-16 07:29  (from work) begin 2 of N (easy, one positional)
        07:36  (done)
        08:36  begin cover help
  04-17 08:12  (from work) help
        16:45  (from Plymouth SW) test docstring thing
        17:52  help screen is covered
	20:25  now