refactor(script-lib): Amazing. Just-in-time parse-parsing ((1708))
Discussion: This was always the plan all along but we bumped it down to
a nice-to-have while we prototyped the engine.

The idea was that in a syntax with several subcommands each with several
options and arguments, we didn't *like* the idea that we were expected
to just accept that we have to parse all of every usage line:

Parsing is in at least two layers, where the usage string is parsed into
a stream of sexps, and each sexp (internally) is converted into a "formal"
object. It felt too clunky for us to require all this work just to land
a request for help on one of several usage lines.

But we had no idea how hard it would be to finally roll this change out,
after we had covered "everything" working without it. As it turned out,
it was just about exactly as hard as we expected.

A requirement of "stops" is that we have two elements of lookahead on the
formal stack (one for the formal at the top of the stack, and one to be
able to look ahead to the next element to see if it is a stop).

To keep it simple we look ahead by two elements at all times, even when
perhaps it isn't necessary.

(This could be made an option trivially to "fail fast" and parse the entire
usage line in all cases.)

(times)
  05-03 00:04  begin experimenting with change interface
               btwn engine and frontend to streaming
  05-03 06:54  (from home) resume
        23:20  (from home) maybe is done, except rules
  05-04 09:18  ok good, after a couple hours