feat(kiss:cap-server): Create entity with form (visually) ((1733))
This commit leaves a lot to be desired (mainly it's not covered)
but because our development took a turn mid-run, we wanted to lock it down
as-is.

We decided to break out redirects into the next run, and have this run
be focused on "flash" messages.

The biggest mid-run change was a slight re-architecting: we discovered
that we really want all the power of the "collection" object in the
html form generation function, not just the formal entity. This botched
the CLI tooling and the unit test, but we recovered in like 20 minutes.

Mid-run (also) we discovered that we again need to lean heavily on
the store-derived abstract entity over the dataclass-derived one.
(The store-derived one knows about primary keys and foreign keys,
which are important dimensions for generating forms.)

The above lead to a reconsideration of what side of attribute names
we wanted to use where. We decided that html forms should use the store-side
names, not the dataclass-side names, because the dataclass-side names are
more verbose and literal and technical to a degree that is not helpful
on the client side.

We have grievances about our experiences trying to use lua that
we will try and save for the project later.

DID:
Solve the encoding problem - see if you can just escape double quotes.
else file
Make the form that is generated be generated, not hand-written
Render the form from formal model (without populating infos of
two categories)
"Flash" messages associated with a particular field (multiple)
Generic "Flash" messages not associated with a particular field
(multiple):
Render the form *with* multiple generic and attribute-focused infos
Visual test of the two kinds of Flash messages
Create button that acts as a link

DID NOT DO:
- Unit test of the two kinds of flash messages
  (to whatever extent is reasonable)
- Does not currently redirect on success
- This writes carriage return characters at the end of multi-line lines
- This does not normalize/validate the rectangles of text

(times)
  05-22 06:23  (from home) Begin looking at form encoding, redirects
        19:47  (from SW Plymouth) begin generate form with infos
  05-23 20:32  (from home) happy with generated form purely from models,
                now try to parse form input from command line with missing
  05-24 09:06  (from GC) hammer out details of repop form with messages
  05-25 07:27  (from GC) did flash messages (over 2 hours)