feat(kiss:cap-server): Discover recsel bug! (visual) ((1733))
(Originally "Bang the hell out of normalizations & encodings")

The reason we categorize this commit as a "feature" is because with this
work in place, it makes it feel like it's starting to get out of the
prototype phase. We can now actually use this to take notes (almost --
we still need to generate ordinals, and also we want the status to be
an editable enum before we traverse the docs.)

Developer's log: making this more robust on the lua side (escaping
the command string we open as a subprocess) lead us to some concern
about potential security risks. Would be better if lua gave us a way
to pass just string tokens.

With our workaround in place (issue an error message), this is beginning
to feel usable.

Also:
   - in .vimrc, don't map <c-v> to anything; we need it to write '\r'

(times)
  05-27 08:48  (from AD) begin a proper, complete normalization of
               "paragraph" type
        14:19  rectangle passed visual test
        17:15  (from shinola) escaping problem
  05-28 07:45  (from Hyperion) fix how we escape strings from lua side
               (very broken)
        14:11  (from AD) spent about an hour or two debugging issue
               with client and backslashes
        17:10  wanted to do a better job covering bug but we would rather
               have the tool first