feat(kiss:cap-server): Create first adaptation of UPDATE for web (visual) ((1733))
Discussion: The big thrilling thing here was re-architecting code between
CREATE and UPDATE and VIEW so that we could re-use view code in our
UPDATE code (with regards to the rendering of html elements). This lead
to an "organic" re-architecting where we started doing crazy things with
attaching properties to element renderers. This would have been smoother
if we had planned for this kind of sharing from the beginning, but who
knows, maybe it would have been over-planned and obtuse.

New Experimental API point in model: `AFTER_CREATE_OR_UPDATE_EXPERIMENTAL`
is how the fent "silo" indicates what page should be displayed after a
successful CREATE or UPDATE. Note this is supposed to be modality agnostic.

LURKING BUG: we aren't translating keys but we should. Somewhere in the
UPDATE code (at least) we're using form parameter names (fpns) and
dataclass attribute names (DANs) interchangeably (as each other) and we
shouldn't be. (The original code was written when there was no
differentiation between DANs and fpns - we used the same keys for both.)

This will definitely manifest as a bug eventually given the right field
names - it's just sheer luck that it hasn't yet. But it's out of scope
until we get to self-supporting-ness and then develop this with coverage.

Also:

- `IDENTIFIER_FOR_PURPOSE` exits experimentality
  (it's enough of a thing now)

(times)
  06-03 20:53  (from home) begin many steps to VIEW
  06-05 08:45  (from AD) (yesterday got form generation done perfectly)
               begin process form for capability UPDATE
  06-06 08:47  (from GC) maybe wire up UPDATE emissions to target specific
               fields. make response structure more interesting
  06-07 09:39  (from Canelle) many form edge cases with empty strings
        09:46  (from SWW) begin cleaning up messy draft of new docpoint