feat(kiss:cap-server): Integrate python Enums (backend) ((1733))
Discussion: beginning to implement the "enum" type effected a pretty
substantial rearchitecting of some of the API: because our use of
dataclass type annotation is a string that's a reference to a different
arbitrary model class, this requires that (in some way) the "whole schema"
is available to any participating modules that may need to resolve such
a class reference.

The pattern is now that we "bake" a function...

This commit is backend-only for "enum"; which means we validate an ordinary
text input against the enum set of allowed string values. In the next
commit we will do the frontend (making it a <select>).

In so doing, we realized that we hadn't put normalization in the UPDATE
yet. We learned other things, like some modality-specific behavior: If you
were running an UPDATE from the command line and one of your provided
values was unchaged from what is stored, we would yes want to express that.
But from a web form, this behavior is unexpected and unhelpful.


(times) (we gave up keeping track after a few days. busy at day job)

  06-07 12:05  begin reading about python enum module
        13:12  (from Avalon) play with enum type as formal type (macro)
               for singular field
  06-10        (did a little stuff from Grenelle (Friday morning))
  06-11        (did a little stuff from AD in morning (Saturday))
  06-12        (did a little stuff from AD 3 times very briefly (Sunday))