feat(kiss-rdb:rec) Begin "tooling" CLI (visual) ((1733))
We do not yet have a "spec" (requirement) for a proper, "formal" CLI
for testing recfile interactions. Indeed it might feel like a misfeature,
since the primary means of interacting with "vanilla" recfiles is thru
its own (quite built-out) CLI.

But we needed a middle-ground between "nothing" and developing everything
visually with "generate_html.py". As such, this "tooling CLI" is born.

There is an awful, old, generic "kiss" CLI for this exact purpose, but
it uses ancient "click" and we don't want to migrate off of that until
"capserver" is complete and abstracted away from its first use case.

See list of all models
  - (needed to expand the thing from `dataclasserer` to `bridger`:
    enter the idea of a "bridge" that transports dataclasses and
    reflection info from the model to the client. Soon we will make
    our production bridge be complicated, but it can be as simple as
    a dict (that has __getitem__ and keys))

See info on one model class
  - Optionally (and experimentally) lets you not load the recfile to see
    the model derived purely from the dataclass

(times)
    07:55  (from Canelle) begin tooling "save"
    10:43  gonna call tooling done for now