data-entry(kiss:cap-server): Begin self-support (date/times)
Discussion: Finally we can get rid of that weirdly-formatted stuff in the
main recfile. (`recins` appears to normalize all of it, file-wide at any
use.)  Feels good that we wrote a single "justification" (in one note)
for the whole section. Taking small-ish steps for now for no real reason.

Despite not having proper nav, this feels sort of usable when we get in
to the rhythm of mostly staying on the keyboard, and opening each click
in a new tab. (But we still want nav)

(times)
  18:18  begin self-support: starting at bottom.
  18:29  now