test(app-flow): Introduce "app flow" ((1737))
Discussion: We did some prototyping and confirmed for ourselves
that we can refactor and abstract our lua end of things to be
reusable across different business domains.

At this point we wanted to (begin to) realize the experiment
of having our app (at least part of it) live as an FSA-like
structure in a database.

To this point, the "UI flow graph" (we need a better name)
lived as a dotfile; but to have the dotfile derived from
a structure in a database as a lot more potential.

Random historical: we consiered different names for the
"Transition" formal entity:
"UI Pathway" "Portal" "Transition" "Arc"

(times)
  06-23 20:38  (from SW Plymouth) It's Time
  06-24 09:23  (from Hyperion) very quick hack on parsing URLs
        09:35  with some prototyping we see that we can do what we want
               (gotta run to work)
        14:45  we want the UI flowchart graph thing
  06-25 08:34  (from AD) gonna try to do this viz quickly
        09:48  (actually start writing CLI)
        16:56  (from SWP) Ok let's really start CLI
        18:03  we have selecting a record and getting a datamodel
        18:40  off to a good start. about to build transitions
        19:40  (from home) resume
        20:04  begin making digraph
        20:24  went well take break
        20:47  resume. gonna cover it
  06-26 06:38  (from home) resume. finish test
        06:49  now