snapshot(app-flow): Snapshot (1/2): We attempted to use a class..
(This is a snapshot of our in-progress work at a moment in time.
Not all tests in the added test file pass. The purpose of this is
merely to show a snapshot of one particular implement: the scanner.

In the next commit we will refactor the scanner to be in a totally
different programming style; and attempt to justify the change
somewhat.)