snapshot(app-flow): Snapshot (2/2): ..But we rewrote it this way
Discussion: This is a case study of doing a typical ordinary object-based
thing a class-based way or our weird way with closures. We know that our
weird way probably irks some readers, so we attempted initially to do it
class-based (the previous commit).

We think these two snapshots offer an interesting exercise in contrasts,
showing the supposed benefits of our closure-based way.

We know *why* we prefer this second way for ourselves (and if we were
writing a blog entry we would elaborate on these reasons, but later for
that. It is our hope that a lot of these reasons are self-evident if you
look at the code in the first way vs the second way. The second way makes
the intent of the implement itself more obvious (it's less code on the
class specifically) ..

BUT ACTUALLY the main reason we wanted to make this snapshot because
we wanted the original two versions "for posterity" so as to do some
benchmarking one day maybe. (The question being, is there significant
overhead to defining a "new" class inside of a clousure like this, and
(if so) does the supposed benefits of this way justify the overhead?)