feat(kiss:cap-server): Integrate "routing" (urls) ((1757))
Discussion: we don't know how far we are going to be able to carry the
vaporware dream of "app flow"; but one essential feature it seemed to
hurt most from was a proper formalization of "routes" (as they call it
in rails).

Also it felt like the biggest "sore thumb" of this as a web app (just
from a look-and-feel standpoint) that we had only one endpoint and a bunch
of GET parameters on every request (most importantly the `action`).

Implementing our bespoke version of "routing" was fun and rewarding,
but also shook things up pretty deeply because of how significant an
architecture change it was.

For reference, these were (more or less) the endpoints we tested visually:

@endpoint('/capability/{EID}/notes/add/')
@endpoint('/capability/{EID}/edit/')
@endpoint('/capability/{EID}/notes/add/', http_method='POST')
@endpoint('/capability/{EID}/edit/', http_method='POST')
@endpoint('/capability/{EID}/')
@endpoint('/')
@endpoint('/', GET_params={'index_style':'tree'})
@endpoint('/test/UI/')
@endpoint('/ping/')

(times)
  07-05 12:15  (from Hyperion) begin integrating routing omg
        20:30  (from SWP) re-arch is heavy around commands and signatures
  07-06 16:52  (from Shinola) ok figured out that our spec isn't that bad
  07-12 07:41  (from Hyperion) we got a form submit working The New Way.
                we just have to finish all the other refactoring
                around url construction
  07-13 14:16  (from Hyperion, home) please let's finish routing
        15:47  (break to go downtown)
        17:51  (from SWW) bash is so annoying to write scripts for
  07-14 19:11  (from SWP) gotta piece together runnable webserver again lol
        20:54  progress was ok. but we need to deal with routes for assets like stylesheets
  07-15 16:48  amazing - served
  07-16 08:15  (from Hyperion) oh noes now that urls have gotten deep,
               assets are loading weird
        08:24  (fixed. was easy. had to introduce absolute urls)
  07-19 11:36  after hours and days and years
        11:37  (from Hyperion) after piecemeal sub-hours after several
               days, maybe we have the story working
        15:48  smoke test & begin assemble commit